% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-oxisols-usda.R
\name{acric_oxisol_usda}
\alias{acric_oxisol_usda}
\title{Acric Oxisol Suborder helper (Acroperox/Acrudox/Acrustox/Acraquox)
Pass when oxic or kandic horizon has ECEC < 1.5 cmol/kg clay AND
pH (KCl) >= 5.0.}
\usage{
acric_oxisol_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Acric Oxisol Suborder helper (Acroperox/Acrudox/Acrustox/Acraquox)
Pass when oxic or kandic horizon has ECEC < 1.5 cmol/kg clay AND
pH (KCl) >= 5.0.
}
