% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-oxisols-usda.R
\name{aeric_oxisol_usda}
\alias{aeric_oxisol_usda}
\title{Aeric Subgroup (for Oxisols Aquox) -- chroma-3 below epipedon
Already defined for Aquods; here we add Oxisol-specific variant
(any 10+ cm horizon below A with chroma >= 3 in 50\%+ peds).}
\usage{
aeric_oxisol_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Aeric Subgroup (for Oxisols Aquox) -- chroma-3 below epipedon
Already defined for Aquods; here we add Oxisol-specific variant
(any 10+ cm horizon below A with chroma >= 3 in 50\%+ peds).
}
