% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-spodosols-usda.R
\name{aeric_subgroup_usda}
\alias{aeric_subgroup_usda}
\title{Aeric Subgroup helper (Aquods)
Pass when ochric epipedon is present (vs. histic/umbric/etc).}
\usage{
aeric_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Aeric Subgroup helper (Aquods)
Pass when ochric epipedon is present (vs. histic/umbric/etc).
}
