% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-ultisols-usda.R
\name{albaquult_qualifying_usda}
\alias{albaquult_qualifying_usda}
\title{Albic-over-argillic qualifying (Albaquults)
Pass when albic horizon overlies an argillic horizon directly.}
\usage{
albaquult_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Albic-over-argillic qualifying (Albaquults)
Pass when albic horizon overlies an argillic horizon directly.
}
