% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-properties-wrb-v033.R
\name{albeluvic_glossae}
\alias{albeluvic_glossae}
\title{Albeluvic glossae (WRB 2022 Ch 3.2.2)}
\usage{
albeluvic_glossae(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Tongues of bleached, coarser-textured material penetrating an argic
horizon. v0.3.3 detects via designation pattern \code{glossic|albeluvic}
on a layer that overlies an argic-horizon-passing layer.
}
