% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-ultisols-usda.R
\name{albic_subgroup_usda}
\alias{albic_subgroup_usda}
\title{Albic Subgroup helper (Albaquultic / Albaquic)}
\usage{
albic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Albic Subgroup helper (Albaquultic / Albaquic)
}
