% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-mollisols-usda.R
\name{alboll_qualifying_usda}
\alias{alboll_qualifying_usda}
\title{Albolls qualifier: mollic + albic + argillic.}
\usage{
alboll_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Albolls qualifier: mollic + albic + argillic.
}
