% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-spodosols-usda.R
\name{alfic_subgroup_usda}
\alias{alfic_subgroup_usda}
\title{Alfic Subgroup helper (Spodosols): argillic or kandic with BS >= 35\%}
\usage{
alfic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Alfic Subgroup helper (Spodosols): argillic or kandic with BS >= 35\%
}
