% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-alfisols-usda.R
\name{alfisol_qualifying_usda}
\alias{alfisol_qualifying_usda}
\title{Alfisol Order qualifier
Pass when argillic OR kandic horizon present + BS >= 35\% in some part.}
\usage{
alfisol_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Alfisol Order qualifier
Pass when argillic OR kandic horizon present + BS >= 35\% in some part.
}
