% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-orders-usda.R
\name{alfisol_usda}
\alias{alfisol_usda}
\title{Alfisols (USDA Cap 5): argillic/kandic/natric horizon + base saturation
>= 35% at the implicit reference depth.}
\usage{
alfisol_usda(pedon)
}
\description{
Alfisols (USDA Cap 5): argillic/kandic/natric horizon + base saturation
>= 35% at the implicit reference depth.
}
