% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-andisols-usda.R
\name{alic_andisol_usda}
\alias{alic_andisol_usda}
\title{Alic Subgroup helper (Andisols)
Pass when al_kcl_cmol > 2.0 in a 10+ cm layer between 25 and 50 cm.}
\usage{
alic_andisol_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Alic Subgroup helper (Andisols)
Pass when al_kcl_cmol > 2.0 in a 10+ cm layer between 25 and 50 cm.
}
