% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-oxisols-usda.R
\name{anionic_subgroup_usda}
\alias{anionic_subgroup_usda}
\title{Anionic Subgroup helper (Oxisols)}
\usage{
anionic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Pass when delta pH (KCl - water) is 0 or net positive in a 18+ cm
layer within 125 cm. Indicates exchange complex dominated by
positive-charge minerals (Fe/Al oxides).
}
