% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-alfisols-usda.R
\name{aqualf_qualifying_usda}
\alias{aqualf_qualifying_usda}
\title{Aqualf Suborder qualifier (aquic conditions in argillic Alfisol).}
\usage{
aqualf_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Aqualf Suborder qualifier (aquic conditions in argillic Alfisol).
}
