% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-andisols-usda.R
\name{aquand_qualifying_usda}
\alias{aquand_qualifying_usda}
\title{Aquands Suborder qualifier (Cap 6, p 117)
Pass when histic OR aquic conditions in 40-50 cm with redox
features. Simplified: histic OR aquic_conditions(max_top=50).}
\usage{
aquand_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Aquands Suborder qualifier (Cap 6, p 117)
Pass when histic OR aquic conditions in 40-50 cm with redox
features. Simplified: histic OR aquic_conditions(max_top=50).
}
