% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-spodosols-usda.R
\name{aquandic_subgroup_usda}
\alias{aquandic_subgroup_usda}
\title{Aquandic Subgroup helper (Spodosols / others)
Aquic + Andic.}
\usage{
aquandic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Aquandic Subgroup helper (Spodosols / others)
Aquic + Andic.
}
