% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-entisols-usda.R
\name{aquent_qualifying_usda}
\alias{aquent_qualifying_usda}
\title{Aquent Suborder qualifier (Entisol with aquic conditions <50 cm).}
\usage{
aquent_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Aquent Suborder qualifier (Entisol with aquic conditions <50 cm).
}
