% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-inceptisols-usda.R
\name{aquept_qualifying_usda}
\alias{aquept_qualifying_usda}
\title{Aquept Suborder qualifier}
\usage{
aquept_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Aquept Suborder qualifier
}
