% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-vertisols-usda.R
\name{aquert_qualifying_usda}
\alias{aquert_qualifying_usda}
\title{Aquerts qualifier (Vertisols with aquic conditions)
Pass when aquic_conditions within 50 cm.}
\usage{
aquert_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Aquerts qualifier (Vertisols with aquic conditions)
Pass when aquic_conditions within 50 cm.
}
