% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-gelisols-usda.R
\name{aquic_subgroup_usda}
\alias{aquic_subgroup_usda}
\title{Aquic Subgroup helper (within 100 cm of mineral soil surface)}
\usage{
aquic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Aquic Subgroup helper (within 100 cm of mineral soil surface)
}
