% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-mollisols-usda.R
\name{aquoll_qualifying_usda}
\alias{aquoll_qualifying_usda}
\title{Aquolls qualifier (aquic conditions in mollic).}
\usage{
aquoll_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Aquolls qualifier (aquic conditions in mollic).
}
