% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-ultisols-usda.R
\name{aquult_qualifying_usda}
\alias{aquult_qualifying_usda}
\title{Aquult Suborder qualifier
Pass when aquic_conditions within 50 cm.}
\usage{
aquult_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Aquult Suborder qualifier
Pass when aquic_conditions within 50 cm.
}
