% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-aridisols-usda.R
\name{argic_aridisol_usda}
\alias{argic_aridisol_usda}
\title{Argic Aridisol helper -- argillic-or-kandic in Argids/Cryids/etc.}
\usage{
argic_aridisol_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Argic Aridisol helper -- argillic-or-kandic in Argids/Cryids/etc.
}
