% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-mollisols-usda.R
\name{argic_mollisol_usda}
\alias{argic_mollisol_usda}
\title{Argic Mollisol Suborder helper -- delegates argillic_within_usda.}
\usage{
argic_mollisol_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Argic Mollisol Suborder helper -- delegates argillic_within_usda.
}
