% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-spodosols-usda.R
\name{argic_subgroup_usda}
\alias{argic_subgroup_usda}
\title{Argic Subgroup helper (Endoaquods/Fragiaquods): argillic or kandic.
Synonym of ultic at this level. Re-exported for naming clarity.}
\usage{
argic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Argic Subgroup helper (Endoaquods/Fragiaquods): argillic or kandic.
Synonym of ultic at this level. Re-exported for naming clarity.
}
