% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{argissolo_acinzentado}
\alias{argissolo_acinzentado}
\title{Argissolos Acinzentados (SiBCS Cap 5)}
\usage{
argissolo_acinzentado(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Matiz \\>= 7.5YR, valor \\>= 5, croma \\< 4 (cores mais cinzentas /
palidas), na maior parte do B (inclusive BA).
}
