% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{argissolo_amarelo}
\alias{argissolo_amarelo}
\title{Argissolos Amarelos (SiBCS Cap 5)}
\usage{
argissolo_amarelo(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Matiz \\>= 7.5YR (mais amarelo) na maior parte do B, sem ser
Acinzentado (croma \\>= 4).
}
