% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{argissolo_bruno_acinzentado}
\alias{argissolo_bruno_acinzentado}
\title{Argissolos Bruno-Acinzentados (SiBCS Cap 5)}
\usage{
argissolo_bruno_acinzentado(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Argissolos com horizonte B textural, com matiz \\>= 5YR e cor escura
(valor \\<= 4 + croma \\<= 4 umidos) na maior parte do B (inclusive
BA).
}
