% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{argissolo_vermelho}
\alias{argissolo_vermelho}
\title{Argissolos Vermelhos (SiBCS Cap 5)}
\usage{
argissolo_vermelho(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Matiz \\<= 2.5YR (mais vermelho) na maior parte do B.
}
