% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-aridisols-usda.R
\name{aridisol_qualifying_usda}
\alias{aridisol_qualifying_usda}
\title{Aridisol Order qualifier (USDA, KST 13ed, Ch 2)
Pass when the soil has aridic SMR AND any one of: argillic, natric,
kandic, calcic, petrocalcic, gypsic, petrogypsic, salic, duripan,
cambic, sulfuric horizon. Also requires no other prior order match.}
\usage{
aridisol_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Aridisol Order qualifier (USDA, KST 13ed, Ch 2)
Pass when the soil has aridic SMR AND any one of: argillic, natric,
kandic, calcic, petrocalcic, gypsic, petrogypsic, salic, duripan,
cambic, sulfuric horizon. Also requires no other prior order match.
}
