% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-orders-usda.R
\name{aridisol_usda}
\alias{aridisol_usda}
\title{Aridisols (USDA Cap 7): aridic moisture regime + ochric/anthropic +
subsurface diagnostic. v0.8 simplification: detected via aridity
proxies (low EC OR salic OR caracter combinations) + non-mollic
surface + low OC (no organic accumulation).}
\usage{
aridisol_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Aridisols (USDA Cap 7): aridic moisture regime + ochric/anthropic +
subsurface diagnostic. v0.8 simplification: detected via aridity
proxies (low EC OR salic OR caracter combinations) + non-mollic
surface + low OC (no organic accumulation).
}
