% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esdb-raster.R
\name{available_esdb_attributes}
\alias{available_esdb_attributes}
\title{List ESDB Raster Library attributes available at a given root}
\usage{
available_esdb_attributes(raster_root)
}
\arguments{
\item{raster_root}{Path to the unpacked ESDB raster directory
(typically `<some>/ESDB-Raster-Library-1k-GeoTIFF-...`).}
}
\value{
A character vector of attribute names (sorted).
}
\description{
Walks `raster_root` and returns the folder names that contain a
valid `<NAME>.tif` raster. Useful for discovery before calling
\code{\link{lookup_esdb}}.
}
\examples{
\dontrun{
available_esdb_attributes("~/data/ESDB-Raster-Library-1k-GeoTIFF-20240507")
#> [1] "AGLI1NNI" "AGLI2NNI" "AGLIM1" "AGLIM2" "ALT" "ATC" "AWC_SUB" ...
#>     [continued: 71 attributes]
}
}
