% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-horizons-sibcs.R
\name{B_espodico}
\alias{B_espodico}
\title{Horizonte B espodico (SiBCS Cap 2, p 62-65; v0.7)}
\usage{
B_espodico(pedon, ...)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{...}{Reserved for future arguments.}
}
\description{
Subsuperficial com acumulo iluvial de Al + Fe + materia organica;
espessura \\>= 2.5 cm. Tipos: Bs, Bhs, Bh, ortstein. Reuso de
\code{\link{spodic}} (WRB) que ja codifica criterios essencialmente
identicos.
}
