% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-gelisols-usda.R
\name{calcic_horizon_usda}
\alias{calcic_horizon_usda}
\title{Calcic horizon (USDA, delegates to WRB calcic).}
\usage{
calcic_horizon_usda(pedon, max_top_cm = 100)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{max_top_cm}{Numeric threshold or option (see Details).}
}
\description{
Calcic horizon (USDA, delegates to WRB calcic).
}
