% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-rsg-sibcs.R
\name{cambissolo}
\alias{cambissolo}
\title{Cambissolos (SiBCS Cap 4, p 113; conceito Cap 3, p 88-89)}
\usage{
cambissolo(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Horizonte B incipiente imediatamente abaixo de A ou histico < 40 cm,
com plintita/petroplintita (se presente) que NAO satisfaca aos
requisitos para Plintossolos.
}
