% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{cambissolo_fluvico}
\alias{cambissolo_fluvico}
\title{Cambissolos Fluvicos (Cap 6): carater fluvico.}
\usage{
cambissolo_fluvico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Cambissolos Fluvicos (Cap 6): carater fluvico.
}
