% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{cambissolo_haplico}
\alias{cambissolo_haplico}
\title{Cambissolos Haplicos (catch-all).}
\usage{
cambissolo_haplico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Cambissolos Haplicos (catch-all).
}
