% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{cambissolo_histico}
\alias{cambissolo_histico}
\title{Cambissolos Histicos (Cap 6): horizonte histico sem espessura para Organossolo.}
\usage{
cambissolo_histico(pedon)
}
\description{
Cambissolos Histicos (Cap 6): horizonte histico sem espessura para Organossolo.
}
