% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{cambissolo_humico}
\alias{cambissolo_humico}
\title{Cambissolos Humicos (Cap 6): horizonte A humico.}
\usage{
cambissolo_humico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Cambissolos Humicos (Cap 6): horizonte A humico.
}
