% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-atributos-sibcs.R
\name{carater_fluvico}
\alias{carater_fluvico}
\title{Carater fluvico (SiBCS Cap 1, p 35-36): camadas estratificadas +
distribuicao irregular de C organico. Reuso de fluvic_material (WRB).}
\usage{
carater_fluvico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Carater fluvico (SiBCS Cap 1, p 35-36): camadas estratificadas +
distribuicao irregular de C organico. Reuso de fluvic_material (WRB).
}
