% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-atributos-sibcs.R
\name{carater_hipocarbonatico}
\alias{carater_hipocarbonatico}
\title{Carater hipocarbonatico (SiBCS Cap 1, p 33): CaCO3 entre 50 e 150 g/kg.}
\usage{
carater_hipocarbonatico(pedon)
}
\description{
Carater hipocarbonatico (SiBCS Cap 1, p 33): CaCO3 entre 50 e 150 g/kg.
}
