% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-atributos-sibcs.R
\name{carater_salico}
\alias{carater_salico}
\title{Carater salico (SiBCS Cap 1, p 38): CE >= 7 dS/m em alguma epoca.}
\usage{
carater_salico(pedon, min_ec = 7)
}
\description{
Carater salico (SiBCS Cap 1, p 38): CE >= 7 dS/m em alguma epoca.
}
