% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-atributos-sibcs.R
\name{carater_solodico}
\alias{carater_solodico}
\title{Carater solodico (SiBCS Cap 1, p 39): PST entre 6% e < 15%.}
\usage{
carater_solodico(pedon, min_pst = 6, max_pst = 15)
}
\description{
Carater solodico (SiBCS Cap 1, p 39): PST entre 6% e < 15%.
}
