% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-rsg-sibcs.R
\name{chernossolo}
\alias{chernossolo}
\title{Chernossolos (SiBCS Cap 4, p 113; conceito Cap 3, p 89-90)}
\usage{
chernossolo(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
A chernozemico seguido de:
(a) Bi OR Bt, ambos com argila ativ alta + V alta; OR
(b) Bi com espessura < 10 cm OR C, ambos calcicos, petrocalcicos
    OR carbonaticos; OR
(c) Calcico OR carater carbonatico no A, seguido de contato
    litico / fragmentario.
}
