% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{chernossolo_argiluvico}
\alias{chernossolo_argiluvico}
\title{Chernossolos Argiluvicos (Cap 7): B textural abaixo do A chernozemico.}
\usage{
chernossolo_argiluvico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Chernossolos Argiluvicos (Cap 7): B textural abaixo do A chernozemico.
}
