% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{chernossolo_ebanico}
\alias{chernossolo_ebanico}
\title{Chernossolos Ebanicos (Cap 7): caracter ebanico em B.
v0.7.1: detecta via Munsell em B - hue 7.5YR ou mais amarelo: V<4 +
C<3 umido; OR hue mais vermelho 7.5YR: preto/cinza muito escuro.}
\usage{
chernossolo_ebanico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Chernossolos Ebanicos (Cap 7): caracter ebanico em B.
v0.7.1: detecta via Munsell em B - hue 7.5YR ou mais amarelo: V<4 +
C<3 umido; OR hue mais vermelho 7.5YR: preto/cinza muito escuro.
}
