% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{chernossolo_haplico}
\alias{chernossolo_haplico}
\title{Chernossolos Haplicos (catch-all).}
\usage{
chernossolo_haplico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Chernossolos Haplicos (catch-all).
}
