% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{chernossolo_rendzico}
\alias{chernossolo_rendzico}
\title{Chernossolos Rendzicos (Cap 7): A chernozemico + (calcico/petrocalcico
OR carater carbonatico).}
\usage{
chernossolo_rendzico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Chernossolos Rendzicos (Cap 7): A chernozemico + (calcico/petrocalcico
OR carater carbonatico).
}
