% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-diagnostic-tests.R
\name{compute_ecec}
\alias{compute_ecec}
\title{Effective CEC from sum of bases plus exchangeable Al}
\usage{
compute_ecec(ca, mg, k, na, al)
}
\description{
If any of \code{ca_cmol}, \code{mg_cmol}, \code{k_cmol}, \code{na_cmol},
\code{al_cmol} are missing, returns \code{NA}.
}
\keyword{internal}
