% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-diagnostic-tests.R
\name{compute_esp}
\alias{compute_esp}
\title{Compute exchangeable sodium percentage (ESP)}
\usage{
compute_esp(na_cmol, cec_cmol)
}
\description{
\code{na_cmol / cec_cmol * 100}, returning NA on missing/zero CEC.
}
\keyword{internal}
