% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-inceptisols-usda.R
\name{densiaquept_qualifying_usda}
\alias{densiaquept_qualifying_usda}
\title{Densiaquept qualifying (densic contact within 100 cm)}
\usage{
densiaquept_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Densiaquept qualifying (densic contact within 100 cm)
}
