% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afsp.R
\name{.afsp_parse_munsell}
\alias{.afsp_parse_munsell}
\title{Parse AfSP Munsell colour string (e.g. "10YR 4/3") into hue/value/chroma}
\usage{
.afsp_parse_munsell(s)
}
\description{
Parse AfSP Munsell colour string (e.g. "10YR 4/3") into hue/value/chroma
}
\keyword{internal}
