% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-rsg-argic-derived.R
\name{.argic_derived_aggregate}
\alias{.argic_derived_aggregate}
\title{Internal helper: .argic_derived_aggregate}
\usage{
.argic_derived_aggregate(tests, layer_keys)
}
\description{
Internal helper: .argic_derived_aggregate
}
\keyword{internal}
