% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-rsg-argic-derived.R
\name{.argic_derived_negative}
\alias{.argic_derived_negative}
\title{Internal helper: .argic_derived_negative}
\usage{
.argic_derived_negative(name, arg_res, note)
}
\description{
Internal helper: .argic_derived_negative
}
\keyword{internal}
