% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdsolos.R
\docType{data}
\name{.BDSOLOS_COLUMN_PATTERNS}
\alias{.BDSOLOS_COLUMN_PATTERNS}
\title{Canonical mapping from BDsolos column-name variants to soilKey schema}
\format{
An object of class \code{list} of length 41.
}
\usage{
.BDSOLOS_COLUMN_PATTERNS
}
\description{
BDsolos exports use Portuguese column names with variable casing and
diacritic handling. This table records the regex patterns that
identify each soilKey horizon column. Patterns are matched
case-insensitively, after stripping diacritics and the underscore
between word fragments.
}
\keyword{internal}
