% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdsolos.R
\name{.bdsolos_dms_to_decimal}
\alias{.bdsolos_dms_to_decimal}
\title{Convert BDsolos coords (graus / minutos / segundos / hemisferio) to decimal}
\usage{
.bdsolos_dms_to_decimal(graus, minutos, segundos, hemisferio)
}
\description{
Convert BDsolos coords (graus / minutos / segundos / hemisferio) to decimal
}
\keyword{internal}
