% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdsolos.R
\name{.bdsolos_eval}
\alias{.bdsolos_eval}
\title{Evaluate JS in a chromote session, returning a string result}
\usage{
.bdsolos_eval(chromote_session, js)
}
\description{
Evaluate JS in a chromote session, returning a string result
}
\keyword{internal}
