% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdsolos.R
\name{.bdsolos_match_taxon_column}
\alias{.bdsolos_match_taxon_column}
\title{Discover taxonomic column (the surveyor's SiBCS classification)}
\usage{
.bdsolos_match_taxon_column(raw_name)
}
\description{
Discover taxonomic column (the surveyor's SiBCS classification)
}
\keyword{internal}
