% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdsolos.R
\name{.bdsolos_rows_to_horizons}
\alias{.bdsolos_rows_to_horizons}
\title{Convert a subset of BDsolos rows (one perfil) to a soilKey horizons table}
\usage{
.bdsolos_rows_to_horizons(rows, sk_map)
}
\description{
Convert a subset of BDsolos rows (one perfil) to a soilKey horizons table
}
\keyword{internal}
