% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdsolos.R
\docType{data}
\name{.BDSOLOS_SITE_PATTERNS}
\alias{.BDSOLOS_SITE_PATTERNS}
\title{Site-level columns (BDsolos full export). Mapped at the site, not
horizon, level.}
\format{
An object of class \code{list} of length 21.
}
\usage{
.BDSOLOS_SITE_PATTERNS
}
\description{
Site-level columns (BDsolos full export). Mapped at the site, not
horizon, level.
}
\keyword{internal}
