% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-unified.R
\name{.benchmark_available_datasets}
\alias{.benchmark_available_datasets}
\title{Which datasets in `paths` actually have their files/dirs present?}
\usage{
.benchmark_available_datasets(paths)
}
\description{
Which datasets in `paths` actually have their files/dirs present?
}
\keyword{internal}
