% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-qgis.R
\name{.build_pedon_point_row}
\alias{.build_pedon_point_row}
\title{Build a single-row tibble describing the profile + classifications
for the GPKG `pedon_point` layer.}
\usage{
.build_pedon_point_row(pedon, classifications, report_html)
}
\description{
Build a single-row tibble describing the profile + classifications
for the GPKG `pedon_point` layer.
}
\keyword{internal}
