% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-pdf.R
\name{.build_report_rmd}
\alias{.build_report_rmd}
\title{Internal helper: .build_report_rmd}
\usage{
.build_report_rmd(results, pedon, title)
}
\description{
Internal helper: .build_report_rmd
}
\keyword{internal}
