% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-diagnostic-tests.R
\name{.candidate_layers}
\alias{.candidate_layers}
\title{Internal helper: .candidate_layers}
\usage{
.candidate_layers(h, candidate_layers = NULL)
}
\description{
Internal helper: .candidate_layers
}
\keyword{internal}
