% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra-predict.R
\name{.check_predict_inputs}
\alias{.check_predict_inputs}
\title{Validate inputs to a prediction backend}
\usage{
.check_predict_inputs(X, properties)
}
\description{
Validate inputs to a prediction backend
}
\keyword{internal}
