% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-atributos-sibcs.R
\name{.classify_decomposition}
\alias{.classify_decomposition}
\title{Classifica grau de decomposicao por camada: saprico / hemico / fibrico}
\usage{
.classify_decomposition(fiber_pct, von_post)
}
\description{
SiBCS Cap 14 adota o criterio USDA Soil Taxonomy:
}
\details{
Saprico:  < 17\% fibras esfregadas  ou  von Post H7-H10
  Hemico:   17-40\% fibras            ou  von Post H5-H6
  Fibrico:  >= 40\% fibras            ou  von Post H1-H4
}
\keyword{internal}
