% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{.coverage_markdown}
\alias{.coverage_markdown}
\title{Render a coverage result as Markdown.}
\usage{
.coverage_markdown(system, res)
}
\description{
Render a coverage result as Markdown.
}
\keyword{internal}
