% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/febr.R
\name{.febr_pedon_from_rows}
\alias{.febr_pedon_from_rows}
\title{Build a single PedonRecord from FEBR rows}
\usage{
.febr_pedon_from_rows(oid, camada_rows, ob_row, ident, hz, ds)
}
\description{
Build a single PedonRecord from FEBR rows
}
\keyword{internal}
