% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize-depths.R
\name{.harmonize_numeric_attrs}
\alias{.harmonize_numeric_attrs}
\title{Numeric attributes via mass-preserving spline}
\usage{
.harmonize_numeric_attrs(hz, attributes, depths, lam)
}
\description{
Numeric attributes via mass-preserving spline
}
\keyword{internal}
