% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-html.R
\name{.html_horizons_table}
\alias{.html_horizons_table}
\title{Render the horizons table from a PedonRecord.}
\usage{
.html_horizons_table(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Render the horizons table from a PedonRecord.
}
\keyword{internal}
