% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-html.R
\name{.html_provenance_table}
\alias{.html_provenance_table}
\title{Render a provenance summary from a PedonRecord.}
\usage{
.html_provenance_table(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Render a provenance summary from a PedonRecord.
}
\keyword{internal}
