% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-html.R
\name{.html_site_header}
\alias{.html_site_header}
\title{Render site metadata header.}
\usage{
.html_site_header(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Render site metadata header.
}
\keyword{internal}
