% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-html.R
\name{.html_summary_table}
\alias{.html_summary_table}
\title{Render the cross-system summary table when multiple results are provided.}
\usage{
.html_summary_table(results)
}
\description{
Render the cross-system summary table when multiple results are provided.
}
\keyword{internal}
