% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical-kst13.R
\name{.kst13_path}
\alias{.kst13_path}
\title{Resolve the path to a vendored canonical KST 13th JSON file}
\usage{
.kst13_path(filename)
}
\description{
Resolve the path to a vendored canonical KST 13th JSON file
}
\keyword{internal}
