% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-lucas-2018.R
\name{.lucas_numeric}
\alias{.lucas_numeric}
\title{Coerce a LUCAS character cell to numeric, treating "< LOD" / "" as NA}
\usage{
.lucas_numeric(x)
}
\description{
Coerce a LUCAS character cell to numeric, treating "< LOD" / "" as NA
}
\keyword{internal}
