% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-performance.R
\name{.make_synth_perf_pedon}
\alias{.make_synth_perf_pedon}
\title{Synthesise a small but realistic 5-horizon pedon for benchmarking}
\usage{
.make_synth_perf_pedon(i)
}
\description{
Synthesise a small but realistic 5-horizon pedon for benchmarking
}
\keyword{internal}
