% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize-depths.R
\name{.modal_by_overlap}
\alias{.modal_by_overlap}
\title{Modal categorical value by depth-overlap fraction}
\usage{
.modal_by_overlap(values, top, bottom, depths)
}
\description{
Modal categorical value by depth-overlap fraction
}
\keyword{internal}
