% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-rsg-mollic-derived.R
\name{.mollic_derived_aggregate}
\alias{.mollic_derived_aggregate}
\title{Internal helper: .mollic_derived_aggregate}
\usage{
.mollic_derived_aggregate(tests, require_pass)
}
\description{
Internal helper: .mollic_derived_aggregate
}
\keyword{internal}
