% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-rsg-mollic-derived.R
\name{.mollic_derived_negative}
\alias{.mollic_derived_negative}
\title{Internal helper: .mollic_derived_negative}
\usage{
.mollic_derived_negative(name, mol_res, note)
}
\description{
Internal helper: .mollic_derived_negative
}
\keyword{internal}
