% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-html.R
\name{.normalise_results}
\alias{.normalise_results}
\title{Internal helper: .normalise_results}
\usage{
.normalise_results(x, pedon = NULL)
}
\description{
Internal helper: .normalise_results
}
\keyword{internal}
