% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/febr.R
\name{.parse_febr_munsell_vec}
\alias{.parse_febr_munsell_vec}
\title{Vectorised Munsell-string parser}
\usage{
.parse_febr_munsell_vec(x)
}
\description{
Returns a data.frame with columns hue / value / chroma, one row per
input string.
}
\keyword{internal}
