% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{.q_weighted_mean}
\alias{.q_weighted_mean}
\title{Volume-weighted mean of a horizon attribute over a depth window}
\usage{
.q_weighted_mean(values, top, bottom, window_top = 0, window_bottom = 100)
}
\description{
Volume-weighted mean of a horizon attribute over a depth window
}
\keyword{internal}
