% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redape.R
\docType{data}
\name{.REDAPE_API_BASE}
\alias{.REDAPE_API_BASE}
\title{Embrapa Redape Dataverse API endpoint}
\format{
An object of class \code{character} of length 1.
}
\usage{
.REDAPE_API_BASE
}
\description{
Embrapa Redape Dataverse API endpoint
}
\keyword{internal}
