% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redape.R
\docType{data}
\name{.REDAPE_GEOTAB_DOI}
\alias{.REDAPE_GEOTAB_DOI}
\title{Default DOI for the Vaz et al. 2023 curated GeoTab dataset}
\format{
An object of class \code{character} of length 1.
}
\usage{
.REDAPE_GEOTAB_DOI
}
\description{
Default DOI for the Vaz et al. 2023 curated GeoTab dataset
}
\keyword{internal}
