% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redape.R
\name{.redape_item_to_pedon}
\alias{.redape_item_to_pedon}
\title{Convert one Redape GeoTab item to a soilKey PedonRecord}
\usage{
.redape_item_to_pedon(item)
}
\description{
Convert one Redape GeoTab item to a soilKey PedonRecord
}
\keyword{internal}
