% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra-neighbours.R
\name{.reduce_for_neighbours}
\alias{.reduce_for_neighbours}
\title{Reduce X (library + query) to a small score space.}
\usage{
.reduce_for_neighbours(X_lib, X_query, y_label)
}
\description{
Reduce X (library + query) to a small score space.
}
\keyword{internal}
