% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-pdf.R
\name{.rmd_classification_block}
\alias{.rmd_classification_block}
\title{Internal helper: .rmd_classification_block}
\usage{
.rmd_classification_block(res)
}
\description{
Internal helper: .rmd_classification_block
}
\keyword{internal}
