% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-pdf.R
\name{.rmd_site_block}
\alias{.rmd_site_block}
\title{Internal helper: .rmd_site_block}
\usage{
.rmd_site_block(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Internal helper: .rmd_site_block
}
\keyword{internal}
