% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil-classes-at-location.R
\name{.rsg_name_for_system}
\alias{.rsg_name_for_system}
\title{Map an RSG code to a human-readable class name in the requested
classification system.}
\usage{
.rsg_name_for_system(code, system)
}
\description{
Map an RSG code to a human-readable class name in the requested
classification system.
}
\keyword{internal}
