% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-smartsolos.R
\name{.smartsolos_clay_films_amt}
\alias{.smartsolos_clay_films_amt}
\title{Map \code{clay_films_amount} (few/common/many) to SmartSolos
\code{CEROSIDADE_QUANTIDADE} (1..3).}
\usage{
.smartsolos_clay_films_amt(x)
}
\description{
Map \code{clay_films_amount} (few/common/many) to SmartSolos
\code{CEROSIDADE_QUANTIDADE} (1..3).
}
\keyword{internal}
