% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-smartsolos.R
\name{.smartsolos_clay_films_strength}
\alias{.smartsolos_clay_films_strength}
\title{Map \code{clay_films_strength} to SmartSolos \code{CEROSIDADE_GRAU} (1..3).}
\usage{
.smartsolos_clay_films_strength(x)
}
\description{
Map \code{clay_films_strength} to SmartSolos \code{CEROSIDADE_GRAU} (1..3).
}
\keyword{internal}
