% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-smartsolos.R
\docType{data}
\name{.SMARTSOLOS_DRAINAGE_SCALE}
\alias{.SMARTSOLOS_DRAINAGE_SCALE}
\title{SmartSolos drainage class scale (DRENAGEM, 1-8)}
\format{
An object of class \code{integer} of length 8.
}
\usage{
.SMARTSOLOS_DRAINAGE_SCALE
}
\description{
SiBCS / Embrapa drainage scale used by the SmartSolosExpert API:
1 excessivamente drenado .. 8 muito mal drenado.
soilKey does not have a canonical drainage column yet; user supplies
via \code{drenagem} argument when known.
}
\keyword{internal}
