% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-smartsolos.R
\name{.smartsolos_response_to_result}
\alias{.smartsolos_response_to_result}
\title{Convert a SmartSolosExpert response to a soilKey ClassificationResult}
\usage{
.smartsolos_response_to_result(resp, pedon, endpoint)
}
\description{
Convert a SmartSolosExpert response to a soilKey ClassificationResult
}
\keyword{internal}
