% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-smartsolos.R
\name{.smartsolos_struct_grade}
\alias{.smartsolos_struct_grade}
\title{Map a soilKey \code{structure_grade} string to the SmartSolos integer
(\code{ESTRUTURA_GRAU}: 1=fraca, 2=moderada, 3=forte).}
\usage{
.smartsolos_struct_grade(x)
}
\description{
Map a soilKey \code{structure_grade} string to the SmartSolos integer
(\code{ESTRUTURA_GRAU}: 1=fraca, 2=moderada, 3=forte).
}
\keyword{internal}
