% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-smartsolos.R
\name{.smartsolos_struct_size}
\alias{.smartsolos_struct_size}
\title{Map \code{structure_size} (very fine .. very coarse) to SmartSolos
\code{ESTRUTURA_TAMANHO} (1..5).}
\usage{
.smartsolos_struct_size(x)
}
\description{
Map \code{structure_size} (very fine .. very coarse) to SmartSolos
\code{ESTRUTURA_TAMANHO} (1..5).
}
\keyword{internal}
