% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-smartsolos.R
\name{.smartsolos_struct_type}
\alias{.smartsolos_struct_type}
\title{Map \code{structure_type} to SmartSolos \code{ESTRUTURA_TIPO} (1..6).}
\usage{
.smartsolos_struct_type(x)
}
\description{
Map \code{structure_type} to SmartSolos \code{ESTRUTURA_TIPO} (1..6).
}
\keyword{internal}
