% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata-lazy-fetch.R
\docType{data}
\name{.SOILKEY_LAZY_FETCH_CACHES}
\alias{.SOILKEY_LAZY_FETCH_CACHES}
\title{Caches managed by the v0.9.94 lazy-fetch system}
\format{
An object of class \code{character} of length 4.
}
\usage{
.SOILKEY_LAZY_FETCH_CACHES
}
\description{
Caches managed by the v0.9.94 lazy-fetch system
}
\keyword{internal}
