% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata-lazy-fetch.R
\docType{data}
\name{.SOILKEY_LAZY_FETCH_RELEASE}
\alias{.SOILKEY_LAZY_FETCH_RELEASE}
\title{Versioned GitHub Release tag where the lazy-fetch caches are pinned}
\format{
An object of class \code{character} of length 1.
}
\usage{
.SOILKEY_LAZY_FETCH_RELEASE
}
\description{
Versioned GitHub Release tag where the lazy-fetch caches are pinned
}
\keyword{internal}
