% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra-fetch-labels.R
\name{.wrb_to_usda_modal_order}
\alias{.wrb_to_usda_modal_order}
\title{Modal USDA Order for a WRB RSG (Schad 2023 Annex Table 1).
Accepts either the singular ("Ferralsol") or plural ("Ferralsols")
form.}
\usage{
.wrb_to_usda_modal_order(rsg)
}
\description{
Modal USDA Order for a WRB RSG (Schad 2023 Annex Table 1).
Accepts either the singular ("Ferralsol") or plural ("Ferralsols")
form.
}
\keyword{internal}
