% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-horizons-sibcs.R
\name{duripa}
\alias{duripa}
\title{Duripa (SiBCS Cap 2, p 74; v0.7)}
\usage{
duripa(pedon, ...)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{...}{Reserved for future arguments.}
}
\description{
Reuso de \code{\link{duric_horizon}} (WRB Ch 3.1): subsuperficial
cimentado por silica, continuo ou em \\>= 50\% volume.
}
