% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-vertisols-usda.R
\name{dystric_subgroup_usda}
\alias{dystric_subgroup_usda}
\title{Dystric Subgroup helper (Vertisols Dystr*)
Pass when BS (NH4OAc) < 50\% in some part of the upper 100 cm.}
\usage{
dystric_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Dystric Subgroup helper (Vertisols Dystr*)
Pass when BS (NH4OAc) < 50\% in some part of the upper 100 cm.
}
