% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-spodosols-usda.R
\name{entic_subgroup_usda}
\alias{entic_subgroup_usda}
\title{Entic Subgroup helper (Spodosols)}
\usage{
entic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Pass when the spodic horizon is "weakly developed":
\itemize{
  \item Less than 1.2\% organic carbon in the upper 10 cm of
        spodic; OR
  \item Spodic horizon < 10 cm thick.
}
}
