% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-orders-usda.R
\name{entisol_usda}
\alias{entisol_usda}
\title{Entisols (USDA Cap 8): catch-all for soils that don't match any
other Order. Always passes.}
\usage{
entisol_usda(pedon)
}
\description{
Entisols (USDA Cap 8): catch-all for soils that don't match any
other Order. Always passes.
}
