% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-spodosols-usda.R
\name{episaturation_usda}
\alias{episaturation_usda}
\title{Episaturation helper (USDA, KST 13ed Ch 3, p 41)
Pass when aquic conditions PLUS perched water (saturation type
"episaturation").}
\usage{
episaturation_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Episaturation helper (USDA, KST 13ed Ch 3, p 41)
Pass when aquic conditions PLUS perched water (saturation type
"episaturation").
}
