% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{espodossolo_ferri_humiluvico}
\alias{espodossolo_ferri_humiluvico}
\title{Espodossolos Ferri-humiluvicos (Cap 8): B espodico tipo Bhs OR
catch-all dos espodossolos.}
\usage{
espodossolo_ferri_humiluvico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Espodossolos Ferri-humiluvicos (Cap 8): B espodico tipo Bhs OR
catch-all dos espodossolos.
}
