% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{espodossolo_ferriluvico}
\alias{espodossolo_ferriluvico}
\title{Espodossolos Ferriluvicos (Cap 8): B espodico tipo Bs (Fe + Al, baixo
OC iluvial).}
\usage{
espodossolo_ferriluvico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Espodossolos Ferriluvicos (Cap 8): B espodico tipo Bs (Fe + Al, baixo
OC iluvial).
}
