% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{espodossolo_humiluvico}
\alias{espodossolo_humiluvico}
\title{Espodossolos Humiluvicos (Cap 8): B espodico tipo Bh (org. + Al,
pouco/sem Fe).}
\usage{
espodossolo_humiluvico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Espodossolos Humiluvicos (Cap 8): B espodico tipo Bh (org. + Al,
pouco/sem Fe).
}
