% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-inceptisols-usda.R
\name{eutric_inceptisol_usda}
\alias{eutric_inceptisol_usda}
\title{Eutric Inceptisol Suborder helper (Eutrudepts)
Pass when BS (NH4OAc) >= 60\% in some part of upper 75 cm.}
\usage{
eutric_inceptisol_usda(pedon, min_bs = 60)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{min_bs}{Numeric threshold or option (see Details).}
}
\description{
Eutric Inceptisol Suborder helper (Eutrudepts)
Pass when BS (NH4OAc) >= 60\% in some part of upper 75 cm.
}
