% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-oxisols-usda.R
\name{eutric_oxisol_usda}
\alias{eutric_oxisol_usda}
\title{Eutric Oxisol Suborder helper (Eutroperox/Eutrudox/etc.)
Pass when BS (NH4OAc) >= 35\% in all layers within 125 cm.}
\usage{
eutric_oxisol_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Eutric Oxisol Suborder helper (Eutroperox/Eutrudox/etc.)
Pass when BS (NH4OAc) >= 35\% in all layers within 125 cm.
}
