% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-andisols-usda.R
\name{eutric_subgroup_usda}
\alias{eutric_subgroup_usda}
\title{Eutric Subgroup helper (Andisols)
Pass when base_saturation (sum-of-cations) >= 50\% in some part.}
\usage{
eutric_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Eutric Subgroup helper (Andisols)
Pass when base_saturation (sum-of-cations) >= 50\% in some part.
}
