% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familia-sibcs.R
\name{familia_grupamento_textural}
\alias{familia_grupamento_textural}
\title{Familia: grupamento textural (Cap 1, p 46)}
\usage{
familia_grupamento_textural(pedon, max_depth_cm = 200)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{max_depth_cm}{Profundidade da secao de controle (default
200 cm).}
}
\value{
\code{\link{FamilyAttribute}}.
}
\description{
Retorna o grupamento textural do solo na secao de controle.
Classes (em g kg-1):
\itemize{
  \item arenosa: areia + areia franca, i.e. (sand_pct - clay_pct) > 70
  \item media: clay < 35 e sand > 15, exceto arenosa
  \item argilosa: clay entre 35 e 60
  \item muito_argilosa: clay > 60
  \item siltosa: clay < 35 e sand < 15
}
}
\details{
Aplicavel a todas as ordens do SiBCS, exceto Neossolos
Quartzarenicos (RQ), nas quais o subgrupamento eh mais
apropriado.
}
\references{
Embrapa (2018), SiBCS 5a ed., Cap 1, p. 46-47;
            Cap 18, p. 281.
}
