% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familia-sibcs.R
\name{familia_label}
\alias{familia_label}
\title{Constroi label textual de Familia a partir de \code{classify_sibcs_familia}}
\usage{
familia_label(familia)
}
\arguments{
\item{familia}{Lista de \code{\link{FamilyAttribute}}, retorno de
\code{\link{classify_sibcs_familia}}.}
}
\value{
String com adjetivos compostos separados por ", ", ou
        vazia se nenhum adjetivo se aplica.
}
\description{
Concatena os \code{value} nao-nulos como string separada por
virgulas, conforme orientado no Cap 18, p 281: "as caracteristicas
utilizadas para identificacao do 5o nivel categorico devem ser
acrescentadas apos a designacao do 4o nivel categorico e separadas
desta e entre si por virgula".
}
