% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familia-sibcs.R
\name{familia_subgrupamento_textural}
\alias{familia_subgrupamento_textural}
\title{Familia: subgrupamento textural (Cap 18, p 283; em validacao)}
\usage{
familia_subgrupamento_textural(pedon, max_depth_cm = 200)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{max_depth_cm}{Profundidade da secao de controle (default
200 cm).}
}
\value{
\code{\link{FamilyAttribute}}.
}
\description{
Subgrupamento textural mais detalhado, aplicavel em
substituicao ao grupamento para Espodossolos, Latossolos
psamiticos, Neossolos Fluvicos Psamiticos, Neossolos
Regoliticos, Neossolos Quartzarenicos, e SGs arenicos /
espessarenicos de Argissolos / Luvissolos / Planossolos /
Plintossolos. Tambem em solos com textura arenosa e/ou media.
}
\details{
Classes (em g kg-1; referidas a media ponderada da secao de controle):
\itemize{
  \item muito_arenosa: classe textural areia (sand >= 85)
  \item arenosa-media: classe textural areia franca (sand >= 70 e
    <= 91; clay <= 15)
  \item media-arenosa: francoarenosa, sand > 52
  \item media-argilosa: franco-argiloarenosa (clay 20-35, sand >=
    45)
  \item media-siltosa: clay < 35 e sand > 15, excluindo as 4
    classes acima
  \item siltosa: clay < 35 e sand < 15
  \item argilosa: clay 35-60
  \item muito_argilosa: clay > 60
}
}
\references{
Embrapa (2018), SiBCS 5a ed., Cap 18, p. 283.
}
