% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familia-usda.R
\name{family_label_usda}
\alias{family_label_usda}
\title{Assemble the USDA family label from family attributes}
\usage{
family_label_usda(family)
}
\arguments{
\item{family}{Named list of \code{\link{FamilyAttribute}}, the return of
\code{\link{classify_usda_family}}.}
}
\value{
Single string (possibly empty).
}
\description{
Joins the non-NULL \code{value}s with ", " in canonical USDA order
(particle-size, mineralogy, CEC-activity, reaction, temperature, depth).
The string is meant to be PREPENDED to the subgroup name.
}
\keyword{internal}
