% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-alfisols-usda.R
\name{ferric_subgroup_usda}
\alias{ferric_subgroup_usda}
\title{Ferric Subgroup helper (Ferrudalfs)
Pass when iron-rich (fe_dcb_pct >= 4\%) horizon present in B.}
\usage{
ferric_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Ferric Subgroup helper (Ferrudalfs)
Pass when iron-rich (fe_dcb_pct >= 4\%) horizon present in B.
}
