% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-histosols-usda.R
\name{fibric_predominant_usda}
\alias{fibric_predominant_usda}
\title{Fibric_predominant_usda: Fibrists Suborder qualifier}
\usage{
fibric_predominant_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Fibric_predominant_usda: Fibrists Suborder qualifier
}
