% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-atributos-sibcs.R
\name{fibrico}
\alias{fibrico}
\title{Material organico fibrico (SiBCS Cap 14)}
\usage{
fibrico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
\code{\link{DiagnosticResult}}.
}
\description{
Material organico pouco decomposto: >= 40\% de fibras esfregadas
OU indice de von Post H1-H4. Discrimina Organossolos Fibricos no
3o nivel.
}
\references{
Embrapa (2018), SiBCS 5a ed., Cap 14 (Organossolos),
            pp 224-226.
}
