% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-entisols-usda.R
\name{fluvent_qualifying_usda}
\alias{fluvent_qualifying_usda}
\title{Fluvent Suborder qualifier (irregular OC decrease in 25-125 cm,
OR layered alluvial designation).}
\usage{
fluvent_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Fluvent Suborder qualifier (irregular OC decrease in 25-125 cm,
OR layered alluvial designation).
}
