% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-gelisols-usda.R
\name{fluventic_usda}
\alias{fluventic_usda}
\title{Fluventic Subgroup helper (irregular OC decrease, NO aquic req.)}
\usage{
fluventic_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Fluventic Subgroup helper (irregular OC decrease, NO aquic req.)
}
