% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-histosols-usda.R
\name{folist_qualifying_usda}
\alias{folist_qualifying_usda}
\title{Folists Suborder qualifier (KST 13ed, Ch 10, p 200)}
\usage{
folist_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Histosols saturated for less than 30 days per year (and not
artificially drained). Implementation: pass when there is no
aquic conditions and no glei designation in the upper 50 cm.
}
