% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-gelisols-usda.R
\name{folistic_subgroup_usda}
\alias{folistic_subgroup_usda}
\title{Folistic Subgroup helper (folistic_epipedon present)}
\usage{
folistic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Folistic Subgroup helper (folistic_epipedon present)
}
