% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-histosols-usda.R
\name{frasic_qualifying_usda}
\alias{frasic_qualifying_usda}
\title{Frasiwassists Subgroup helper (Wassists)}
\usage{
frasic_qualifying_usda(pedon, max_ec = 0.6, max_top_cm = 100)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{max_ec}{Default 0.6.}

\item{max_top_cm}{Default 100.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Pass when ec_dS_m < 0.6 (1:5 soil:water) in all horizons within
100 cm. KST 13ed, Ch 10, p 203.
}
