% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{gleissolo_haplico}
\alias{gleissolo_haplico}
\title{Gleissolos Haplicos (catch-all).}
\usage{
gleissolo_haplico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Gleissolos Haplicos (catch-all).
}
