% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{gleissolo_melanico}
\alias{gleissolo_melanico}
\title{Gleissolos Melanicos (Cap 9): horizonte hístico < 40 cm OR A humico,
proeminente, chernozemico.}
\usage{
gleissolo_melanico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Gleissolos Melanicos (Cap 9): horizonte hístico < 40 cm OR A humico,
proeminente, chernozemico.
}
