% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{gleissolo_salico}
\alias{gleissolo_salico}
\title{Gleissolos Salicos (Cap 9): caracter salico em < 100 cm.}
\usage{
gleissolo_salico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Gleissolos Salicos (Cap 9): caracter salico em < 100 cm.
}
