% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{gleissolo_tiomorfico}
\alias{gleissolo_tiomorfico}
\title{Gleissolos Tiomorficos (Cap 9): materiais sulfidricos OR horizonte
sulfurico em < 100 cm.}
\usage{
gleissolo_tiomorfico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Gleissolos Tiomorficos (Cap 9): materiais sulfidricos OR horizonte
sulfurico em < 100 cm.
}
