% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-alfisols-usda.R
\name{glossic_subgroup_usda}
\alias{glossic_subgroup_usda}
\title{Glossic Subgroup helper (Glossaqualfs, Glossocryalfs, Glossudalfs)
Pass when interfingering of albic materials into argillic horizon
is detected. v0.8 proxy: albic + argillic + lateral chroma <= 2
on argillic boundary.}
\usage{
glossic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Glossic Subgroup helper (Glossaqualfs, Glossocryalfs, Glossudalfs)
Pass when interfingering of albic materials into argillic horizon
is detected. v0.8 proxy: albic + argillic + lateral chroma <= 2
on argillic boundary.
}
