% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grapes-or-or-grapes}
\alias{grapes-or-or-grapes}
\alias{\%||\%}
\title{Default-value-for-NULL operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{The candidate value.}

\item{b}{The fallback used when \code{a} is NULL.}
}
\value{
Either \code{a} or \code{b}.
}
\description{
Returns the left-hand side if it is non-NULL, otherwise the right-hand side.
Re-exported so that downstream code can use the same idiom soilKey itself
uses internally.
}
