% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-spodosols-usda.R
\name{grossarenic_subgroup_usda}
\alias{grossarenic_subgroup_usda}
\title{Grossarenic Subgroup helper: sandy throughout, spodic >= 125 cm.}
\usage{
grossarenic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Grossarenic Subgroup helper: sandy throughout, spodic >= 125 cm.
}
