% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize-depths.R
\docType{data}
\name{GSM_DEPTHS}
\alias{GSM_DEPTHS}
\title{Default GlobalSoilMap depth intervals (cm)}
\format{
An object of class \code{numeric} of length 7.
}
\usage{
GSM_DEPTHS
}
\description{
GSM standard per Arrouays et al. (2014) "GlobalSoilMap: Toward a
fine-resolution global grid of soil properties". Boundaries:
0-5, 5-15, 15-30, 30-60, 60-100, 100-200 cm.
}
\keyword{datasets}
