% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-inceptisols-usda.R
\name{halaquept_qualifying_usda}
\alias{halaquept_qualifying_usda}
\title{Halic helper for Halaquepts
Pass when EC >= 8 dS/m within 100 cm.}
\usage{
halaquept_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Halic helper for Halaquepts
Pass when EC >= 8 dS/m within 100 cm.
}
