% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-histosols-usda.R
\name{hemic_subgroup_usda}
\alias{hemic_subgroup_usda}
\title{Hemic Subgroup helper}
\usage{
hemic_subgroup_usda(pedon, max_top_cm = 130)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{max_top_cm}{Numeric threshold or option (see Details).}
}
\description{
Hemic Subgroup helper
}
