% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-atributos-sibcs.R
\name{hemico}
\alias{hemico}
\title{Material organico hemico (SiBCS Cap 14)}
\usage{
hemico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
\code{\link{DiagnosticResult}}.
}
\description{
Material organico em decomposicao intermediaria: 17-40\% de fibras
esfregadas OU indice de von Post H5-H6. Discrimina Organossolos
Hemicos no 3o nivel.
}
\references{
Embrapa (2018), SiBCS 5a ed., Cap 14 (Organossolos),
            pp 224-226.
}
