% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-gelisols-usda.R
\name{histel_qualifying_usda}
\alias{histel_qualifying_usda}
\title{Histels Suborder qualifier (USDA, KST 13ed)}
\usage{
histel_qualifying_usda(pedon, min_thickness_cm = 40, max_top_cm = 50)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{min_thickness_cm}{Default 40.}

\item{max_top_cm}{Default 50.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Pass when a Gelisol has organic soil materials that:
\itemize{
  \item Total >= 40 cm cumulative thickness within 0-50 cm; OR
  \item Comprise >= 80\% (by volume) of 0-50 cm.
}
KST 13ed, Ch 9, p 189 (item AA in Key to Suborders).
}
\references{
Soil Survey Staff (2022), KST 13ed, Ch. 9, p 189.
}
