% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-spodosols-usda.R
\name{histic_subgroup_usda}
\alias{histic_subgroup_usda}
\title{Histic Subgroup helper (in Spodosols, Aquods)
Pass when histic_epipedon_usda passes.}
\usage{
histic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Histic Subgroup helper (in Spodosols, Aquods)
Pass when histic_epipedon_usda passes.
}
