% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-horizons-sibcs.R
\name{horizonte_A_antropico}
\alias{horizonte_A_antropico}
\title{Horizonte A antropico (SiBCS) (SiBCS Cap 2, p 53)}
\usage{
horizonte_A_antropico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Antropic surface formed by long human use; \\>= 20 cm + P Mehlich-1
\\>= 30 mg/kg + evidencias antropogenicas. Reuso de \code{\link{hortic}}
(WRB) com criterios SiBCS-specific.
}
