% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-horizons-sibcs.R
\name{horizonte_A_fraco}
\alias{horizonte_A_fraco}
\title{Horizonte A fraco (SiBCS Cap 2, p 53): cor clara + estrutura grao
simples/macica + OC < 6 g/kg; OR espessura < 5 cm.}
\usage{
horizonte_A_fraco(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Horizonte A fraco (SiBCS Cap 2, p 53): cor clara + estrutura grao
simples/macica + OC < 6 g/kg; OR espessura < 5 cm.
}
