% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-horizons-sibcs.R
\name{horizonte_calcico}
\alias{horizonte_calcico}
\title{Horizonte calcico (SiBCS Cap 2, p 71-72; v0.7)}
\usage{
horizonte_calcico(pedon, ...)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{...}{Reserved for future arguments.}
}
\description{
Reuso direto de \code{\link{calcic}} (WRB Ch 3.1.5) -- criterios
identicos: 150 g/kg CaCO3 + 50 g/kg a mais que sub-jacente +
espessura \\>= 15 cm.
}
