% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-horizons-sibcs.R
\name{horizonte_E_albico}
\alias{horizonte_E_albico}
\title{Horizonte E albico (SiBCS Cap 2, p 66-67; v0.7)}
\usage{
horizonte_E_albico(pedon, ...)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{...}{Reserved for future arguments.}
}
\description{
Reuso de \code{\link{albic}} (WRB Ch 3.1) com criterios identicos.
}
