% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-horizons-sibcs.R
\name{horizonte_petrocalcico}
\alias{horizonte_petrocalcico}
\title{Horizonte petrocalcico (SiBCS Cap 2, p 72; v0.7)}
\usage{
horizonte_petrocalcico(pedon, ...)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{...}{Reserved for future arguments.}
}
\description{
Reuso de \code{\link{petrocalcic}} (WRB v0.3.3).
}
