% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-horizons-sibcs.R
\name{horizonte_sulfurico}
\alias{horizonte_sulfurico}
\title{Horizonte sulfurico (SiBCS Cap 2, p 72-73; v0.7)}
\usage{
horizonte_sulfurico(pedon, ...)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{...}{Reserved for future arguments.}
}
\description{
Reuso de \code{\link{thionic}} (WRB v0.3.3): pH \\<= 3.5 + sulfidic
material + espessura \\>= 15 cm.
}
