% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-andisols-usda.R
\name{humic_andisol_usda}
\alias{humic_andisol_usda}
\title{Humic Andisols Subgroup helper
Pass when mollic OR umbric epipedon present.}
\usage{
humic_andisol_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Humic Andisols Subgroup helper
Pass when mollic OR umbric epipedon present.
}
