% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-spodosols-usda.R
\name{humic_spodic_usda}
\alias{humic_spodic_usda}
\title{Humic-spodic Suborder/GG check (>= 6\% OC in 10+ cm of spodic)}
\usage{
humic_spodic_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Humic-spodic Suborder/GG check (>= 6\% OC in 10+ cm of spodic)
}
