% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-spodosols-usda.R
\name{humic_subgroup_usda}
\alias{humic_subgroup_usda}
\title{Humic Subgroup helper (Humic Duricryods / Humic Placocryods)
Pass when spodic horizon has >= 6\% OC in 10+ cm.}
\usage{
humic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Humic Subgroup helper (Humic Duricryods / Humic Placocryods)
Pass when spodic horizon has >= 6\% OC in 10+ cm.
}
