% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-entisols-usda.R
\name{hydraquent_qualifying_usda}
\alias{hydraquent_qualifying_usda}
\title{Hydric Aquent helper (Hydraquents)
Pass when surface 0-50 has high water content (n value high).
v0.8 proxy: water_content_1500kpa >= 80\% in surface.}
\usage{
hydraquent_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Hydric Aquent helper (Hydraquents)
Pass when surface 0-50 has high water content (n value high).
v0.8 proxy: water_content_1500kpa >= 80\% in surface.
}
