% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-andisols-usda.R
\name{hydric_andisol_usda}
\alias{hydric_andisol_usda}
\title{Hydric (Andisols): 1500 kPa water retention >= 70\% on undried
samples throughout a 35+ cm layer within 100 cm.}
\usage{
hydric_andisol_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Hydric (Andisols): 1500 kPa water retention >= 70\% on undried
samples throughout a 35+ cm layer within 100 cm.
}
