% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-histosols-usda.R
\name{hydric_subgroup_usda}
\alias{hydric_subgroup_usda}
\title{Hydric Subgroup helper (Histosols Cryofibrists / Sphagnofibrists /
etc.)}
\usage{
hydric_subgroup_usda(pedon, max_top_cm = 130)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{max_top_cm}{Default 130.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Pass when there is a "layer of water" within the control section.
Detected via designation containing "W" (water layer) or
\code{layer_origin == "water"}.
}
