% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-inceptisols-usda.R
\name{inceptisol_qualifying_usda}
\alias{inceptisol_qualifying_usda}
\title{Inceptisol Order qualifier
Pass when a cambic horizon is present (no argillic, no spodic,
no mollic, etc. -- enforced by prior order exclusion).}
\usage{
inceptisol_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Inceptisol Order qualifier
Pass when a cambic horizon is present (no argillic, no spodic,
no mollic, etc. -- enforced by prior order exclusion).
}
