% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-orders-usda.R
\name{inceptisol_usda}
\alias{inceptisol_usda}
\title{Inceptisols (USDA Cap 11): cambic horizon (or several alternative
subsurface diagnostics: folistic/histic/mollic with thin sub, salic,
sodium-affected sub).}
\usage{
inceptisol_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Inceptisols (USDA Cap 11): cambic horizon (or several alternative
subsurface diagnostics: folistic/histic/mollic with thin sub, salic,
sodium-affected sub).
}
