% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-oxisols-usda.R
\name{kandic_oxisol_usda}
\alias{kandic_oxisol_usda}
\title{Kandic Suborder helper for Oxisols (Kandiperox/Kandiudox/Kandiustox)
Delegates to kandic_horizon_usda.}
\usage{
kandic_oxisol_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Kandic Suborder helper for Oxisols (Kandiperox/Kandiudox/Kandiustox)
Delegates to kandic_horizon_usda.
}
