% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-ultisols-usda.R
\name{kanhapl_qualifying_usda}
\alias{kanhapl_qualifying_usda}
\title{Kanhapl qualifying helper (Kanhapludults / Kanhaplustults / etc.)
Pass when kandic horizon present BUT NOT meeting Pale criteria
(i.e. younger / less developed kandic).}
\usage{
kanhapl_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Kanhapl qualifying helper (Kanhapludults / Kanhaplustults / etc.)
Pass when kandic horizon present BUT NOT meeting Pale criteria
(i.e. younger / less developed kandic).
}
