% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{latossolo_amarelo}
\alias{latossolo_amarelo}
\title{Latossolos Amarelos (Cap 10): matiz \\>= 7.5YR (mais amarelo).}
\usage{
latossolo_amarelo(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Latossolos Amarelos (Cap 10): matiz \\>= 7.5YR (mais amarelo).
}
