% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{latossolo_bruno}
\alias{latossolo_bruno}
\title{Latossolos Brunos (Cap 10): matiz \\>= 7.5YR + valor \\<= 4 + croma
\\<= 5 (cores brunadas) OR caracter retratil.}
\usage{
latossolo_bruno(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Latossolos Brunos (Cap 10): matiz \\>= 7.5YR + valor \\<= 4 + croma
\\<= 5 (cores brunadas) OR caracter retratil.
}
