% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{latossolo_vermelho}
\alias{latossolo_vermelho}
\title{Latossolos Vermelhos (Cap 10): matiz \\<= 2.5YR (mais vermelho).}
\usage{
latossolo_vermelho(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Latossolos Vermelhos (Cap 10): matiz \\<= 2.5YR (mais vermelho).
}
