% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{latossolo_vermelho_amarelo}
\alias{latossolo_vermelho_amarelo}
\title{Latossolos Vermelho-Amarelos (catch-all).}
\usage{
latossolo_vermelho_amarelo(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Latossolos Vermelho-Amarelos (catch-all).
}
