% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-rsg-sibcs.R
\name{luvissolo}
\alias{luvissolo}
\title{Luvissolos (SiBCS Cap 4, p 113; conceito Cap 3, p 95-96)}
\usage{
luvissolo(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Horizonte B textural com argila ativ alta E saturacao por bases
alta (V \\>= 50\%) na maior parte dos primeiros 100 cm do B
(incl. BA), abaixo de A ou E.
}
