% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{luvissolo_cromico}
\alias{luvissolo_cromico}
\title{Luvissolos Cromicos (Cap 11): caracter cromico (cores fortes em B).
Aplicado pela presenca de Munsell vermelho-amarelado em B com cromas
altos.}
\usage{
luvissolo_cromico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Luvissolos Cromicos (Cap 11): caracter cromico (cores fortes em B).
Aplicado pela presenca de Munsell vermelho-amarelado em B com cromas
altos.
}
