% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{luvissolo_haplico}
\alias{luvissolo_haplico}
\title{Luvissolos Haplicos (catch-all).}
\usage{
luvissolo_haplico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Luvissolos Haplicos (catch-all).
}
