% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fixtures-sibcs.R
\name{make_argissolo_canonical}
\alias{make_argissolo_canonical}
\title{Perfil canonico de Argissolo (SiBCS 5a ed., Cap 5)}
\usage{
make_argissolo_canonical()
}
\description{
B textural com gradiente significativo, argila ativ baixa ou
alta + V baixa. Catch-all final na chave -- tipica do Brasil
tropical.
}
