% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fixtures-sibcs.R
\name{make_luvissolo_canonical}
\alias{make_luvissolo_canonical}
\title{Perfil canonico de Luvissolo (SiBCS 5a ed., Cap 11)}
\usage{
make_luvissolo_canonical()
}
\description{
Solo com B textural argila Ta + V alta. Tipico do semiarido com
rocha basica.
}
