% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fixtures-sibcs.R
\name{make_plintossolo_canonical}
\alias{make_plintossolo_canonical}
\title{Perfil canonico de Plintossolo (SiBCS 5a ed., Cap 16)}
\usage{
make_plintossolo_canonical()
}
\description{
Reusa fixture WRB Plinthosol -- horizonte plintico iniciando
dentro de 40 cm.
}
