% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-andisols-usda.R
\name{melanic_andisol_usda}
\alias{melanic_andisol_usda}
\title{Melanic Andisols: melanic_epipedon present.}
\usage{
melanic_andisol_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Melanic Andisols: melanic_epipedon present.
}
