% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-mollisols-usda.R
\name{mollisol_qualifying_usda}
\alias{mollisol_qualifying_usda}
\title{Mollisol Order qualifier (USDA, KST 13ed, Ch 12)
Pass when mollic_epipedon AND BS (NH4OAc) >= 50\% in upper 100 cm.}
\usage{
mollisol_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Mollisol Order qualifier (USDA, KST 13ed, Ch 12)
Pass when mollic_epipedon AND BS (NH4OAc) >= 50\% in upper 100 cm.
}
