% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-atributos-sibcs.R
\name{mudanca_textural_abrupta}
\alias{mudanca_textural_abrupta}
\title{Mudanca textural abrupta (SiBCS Cap 1, p 30-31)}
\usage{
mudanca_textural_abrupta(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Aumento consideravel de argila em pequena distancia vertical
(\\<= 7.5 cm) na transicao A/E -> B:
\itemize{
  \item argila A < 200 g/kg: argila B \\>= 2x A; OR
  \item argila A 200-400 g/kg: incremento absoluto \\>= 200 g/kg
        (i.e. de 300 -> 500); OR
  \item argila A \\>= 400 g/kg: incremento absoluto \\>= 220 g/kg
        (i.e. de 420 -> 640).
}
Reuso de \code{\link{abrupt_textural_difference}} (WRB Ch 3.2.1)
que ja codifica criterios essencialmente equivalentes.
}
