% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-vertisols-usda.R
\name{natric_subgroup_usda}
\alias{natric_subgroup_usda}
\title{Natric Subgroup helper for Natraquerts.}
\usage{
natric_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Natric Subgroup helper for Natraquerts.
}
