% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-rsg-sibcs.R
\name{neossolo}
\alias{neossolo}
\title{Neossolos (SiBCS Cap 4, p 111-112; conceito Cap 3, p 96-97)}
\usage{
neossolo(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Solos pouco evoluidos: SEM horizonte B diagnostico + ausencia de:
(a) glei dentro 150 cm, (b) plintico dentro 40 cm, (c) vertico
imediatamente abaixo de A, (d) A chernozemico conjugado com
carbonatico ou cálcico.
}
