% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{neossolo_fluvico}
\alias{neossolo_fluvico}
\title{Neossolos Fluvicos (Cap 12): caracter fluvico em < 150 cm.}
\usage{
neossolo_fluvico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Neossolos Fluvicos (Cap 12): caracter fluvico em < 150 cm.
}
