% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{neossolo_regolitico}
\alias{neossolo_regolitico}
\title{Neossolos Regoliticos (catch-all dos Neossolos).}
\usage{
neossolo_regolitico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Neossolos Regoliticos (catch-all dos Neossolos).
}
