% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-rsg-sibcs.R
\name{nitossolo}
\alias{nitossolo}
\title{Nitossolos (SiBCS Cap 4, p 114; conceito Cap 3, p 97-98)}
\usage{
nitossolo(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
\\>= 350 g/kg argila incluindo no horizonte A, com B nitico abaixo
do A, com argila ativ baixa OR ativ alta + carater alumínico, na
maior parte dos primeiros 100 cm do B (incl. BA).
}
