% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{nitossolo_haplico}
\alias{nitossolo_haplico}
\title{Nitossolos Haplicos (catch-all).}
\usage{
nitossolo_haplico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Nitossolos Haplicos (catch-all).
}
