% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{nitossolo_vermelho}
\alias{nitossolo_vermelho}
\title{Nitossolos Vermelhos (Cap 13): matiz \\<= 2.5YR.}
\usage{
nitossolo_vermelho(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Nitossolos Vermelhos (Cap 13): matiz \\<= 2.5YR.
}
