% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{organossolo_haplico}
\alias{organossolo_haplico}
\title{Organossolos Haplicos (catch-all).}
\usage{
organossolo_haplico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Organossolos Haplicos (catch-all).
}
