% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{organossolo_tiomorfico}
\alias{organossolo_tiomorfico}
\title{Organossolos Tiomorficos (Cap 14): materiais sulfidricos OR
horizonte sulfurico em < 100 cm.}
\usage{
organossolo_tiomorfico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Organossolos Tiomorficos (Cap 14): materiais sulfidricos OR
horizonte sulfurico em < 100 cm.
}
