% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-oxisols-usda.R
\name{oxic_horizon_usda}
\alias{oxic_horizon_usda}
\title{Oxic horizon (USDA, KST 13ed, Ch 3)
Delegates to WRB \code{ferralic}.}
\usage{
oxic_horizon_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Oxic horizon (USDA, KST 13ed, Ch 3)
Delegates to WRB \code{ferralic}.
}
