% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-andisols-usda.R
\name{pachic_subgroup_usda}
\alias{pachic_subgroup_usda}
\title{Pachic Subgroup helper (Andisols, Mollisols)
Pass when mollic OR umbric epipedon is >= 50 cm thick.}
\usage{
pachic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Pachic Subgroup helper (Andisols, Mollisols)
Pass when mollic OR umbric epipedon is >= 50 cm thick.
}
