% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-ultisols-usda.R
\name{pale_qualifying_usda}
\alias{pale_qualifying_usda}
\title{Pale qualifying helper (Paleudults / Paleustults / Palexerults /
Palehumults / Paleaquults)}
\usage{
pale_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Pass when an argillic horizon has either:
\itemize{
  \item clay >= 35\% in upper 30 cm of argillic; OR
  \item lithologic discontinuity NOT followed by argic; OR
  \item argillic that does NOT decrease in clay >= 20\% relative
        from its maximum.
}
v0.8 proxy: clay_pct >= 35\% in upper argillic.
}
