% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-aridisols-usda.R
\name{petrocalcic_subgroup_usda}
\alias{petrocalcic_subgroup_usda}
\title{Petrocalcic Subgroup helper (Aridisols Petrocalcids)
Cemented calcic horizon with cementation_class >= "strongly".}
\usage{
petrocalcic_subgroup_usda(pedon, max_top_cm = 100)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{max_top_cm}{Numeric threshold or option (see Details).}
}
\description{
Petrocalcic Subgroup helper (Aridisols Petrocalcids)
Cemented calcic horizon with cementation_class >= "strongly".
}
