% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-aridisols-usda.R
\name{petrogypsic_subgroup_usda}
\alias{petrogypsic_subgroup_usda}
\title{Petrogypsic Subgroup helper -- delegate to petrogypsic_horizon_usda}
\usage{
petrogypsic_subgroup_usda(pedon, max_top_cm = 100)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{max_top_cm}{Numeric threshold or option (see Details).}
}
\description{
Petrogypsic Subgroup helper -- delegate to petrogypsic_horizon_usda
}
