% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-rsg-sibcs.R
\name{planossolo}
\alias{planossolo}
\title{Planossolos (SiBCS Cap 4, p 112; conceito Cap 3, p 101-102)}
\usage{
planossolo(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Horizonte B planico nao coincidente com plintico (sem carater
sodico), imediatamente abaixo de A ou E.
}
