% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{planossolo_haplico}
\alias{planossolo_haplico}
\title{Planossolos Haplicos (catch-all).}
\usage{
planossolo_haplico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Planossolos Haplicos (catch-all).
}
