% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{planossolo_natrico}
\alias{planossolo_natrico}
\title{Planossolos Natricos (Cap 15): caracter sodico em \\< 100 cm.}
\usage{
planossolo_natrico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Planossolos Natricos (Cap 15): caracter sodico em \\< 100 cm.
}
