% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-ultisols-usda.R
\name{plinth_subgroup_usda}
\alias{plinth_subgroup_usda}
\title{Plinth qualifying helper (Plinth*ults)
Pass when plinthite >= 5\% in 50\%+ of layers within 150 cm.}
\usage{
plinth_subgroup_usda(pedon, max_top_cm = 150)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{max_top_cm}{Numeric threshold or option (see Details).}
}
\description{
Plinth qualifying helper (Plinth*ults)
Pass when plinthite >= 5\% in 50\%+ of layers within 150 cm.
}
