% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-oxisols-usda.R
\name{plinthaquox_qualifying_usda}
\alias{plinthaquox_qualifying_usda}
\title{Plinthaquox qualifying helper (Aquox: continuous plinthite phase)
Pass when plinthite >= 50\% in some 10+ cm layer (continuous phase
proxy).}
\usage{
plinthaquox_qualifying_usda(pedon, max_top_cm = 125)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{max_top_cm}{Numeric threshold or option (see Details).}
}
\description{
Plinthaquox qualifying helper (Aquox: continuous plinthite phase)
Pass when plinthite >= 50\% in some 10+ cm layer (continuous phase
proxy).
}
