% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-oxisols-usda.R
\name{plinthic_subgroup_usda}
\alias{plinthic_subgroup_usda}
\title{Plinthic Subgroup helper (Oxisols)
Pass when plinthite >= 5\% in any horizon within 125 cm.}
\usage{
plinthic_subgroup_usda(pedon, max_top_cm = 125)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{max_top_cm}{Numeric threshold or option (see Details).}
}
\description{
Plinthic Subgroup helper (Oxisols)
Pass when plinthite >= 5\% in any horizon within 125 cm.
}
