% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-rsg-sibcs.R
\name{plintossolo}
\alias{plintossolo}
\title{Plintossolos (SiBCS Cap 4, p 113; conceito Cap 3, p 102-104)}
\usage{
plintossolo(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Horizonte plintico (nao coincidente com B planico de carater sodico),
OR litoplintico, OR concrecionario, iniciando dentro de 40 cm OR
dentro de 200 cm precedido de glei OR A/E OR horizonte com cores
palidas / variegadas / mosqueados.
}
