% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{plintossolo_argiluvico}
\alias{plintossolo_argiluvico}
\title{Plintossolos Argiluvicos (Cap 16): horizonte plintico + B textural OR
carater argiluvico.}
\usage{
plintossolo_argiluvico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Plintossolos Argiluvicos (Cap 16): horizonte plintico + B textural OR
carater argiluvico.
}
