% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{plintossolo_haplico}
\alias{plintossolo_haplico}
\title{Plintossolos Haplicos (catch-all).}
\usage{
plintossolo_haplico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Plintossolos Haplicos (catch-all).
}
