% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{plintossolo_petrico}
\alias{plintossolo_petrico}
\title{Plintossolos Petricos (Cap 16): horizonte concrecionario OR
litoplintico (sem horizonte plintico precedendo).}
\usage{
plintossolo_petrico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Plintossolos Petricos (Cap 16): horizonte concrecionario OR
litoplintico (sem horizonte plintico precedendo).
}
