% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra-train.R
\name{print.soilKey_pls_model}
\alias{print.soilKey_pls_model}
\title{Print method for soilKey_pls_model}
\usage{
\method{print}{soilKey_pls_model}(x, ...)
}
\arguments{
\item{x}{A \code{soilKey_pls_model} object.}

\item{...}{Reserved.}
}
\value{
The object, invisibly.
}
\description{
Print method for soilKey_pls_model
}
