% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-entisols-usda.R
\name{psamment_qualifying_usda}
\alias{psamment_qualifying_usda}
\title{Psamment Suborder qualifier (sandy texture: clay + 2*silt < 30
AND no clay films / argillic).}
\usage{
psamment_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Psamment Suborder qualifier (sandy texture: clay + 2*silt < 30
AND no clay films / argillic).
}
