% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-gelisols-usda.R
\name{psammentic_subgroup_usda}
\alias{psammentic_subgroup_usda}
\title{Psammentic Subgroup helper (Aquorthels)}
\usage{
psammentic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Pass when, in particle-size control section: < 35\% rock
fragments AND texture class loamy fine sand or coarser in all
layers.
}
