% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091b.R
\name{qual_abruptic}
\alias{qual_abruptic}
\title{Abruptic qualifier (ap): abrupt textural difference within 100 cm.}
\usage{
qual_abruptic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Abruptic qualifier (ap): abrupt textural difference within 100 cm.
}
