% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_acric}
\alias{qual_acric}
\title{Acric qualifier (ac): argic horizon + low CEC + high Al.
v0.9: argic + CEC < 24 cmolc/kg clay + exch Al > Ca+Mg+K+Na.}
\usage{
qual_acric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Acric qualifier (ac): argic horizon + low CEC + high Al.
v0.9: argic + CEC < 24 cmolc/kg clay + exch Al > Ca+Mg+K+Na.
}
