% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091b.R
\name{qual_acroxic}
\alias{qual_acroxic}
\title{Acroxic qualifier (ax): andic + extremely low effective exchange
complex (Ca + Mg + K + Na exch + 1 N KCl Al-exch <= 2 cmol+/kg fine
earth) in some layer of the andic part within 100 cm.}
\usage{
qual_acroxic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Acroxic qualifier (ax): andic + extremely low effective exchange
complex (Ca + Mg + K + Na exch + 1 N KCl Al-exch <= 2 cmol+/kg fine
earth) in some layer of the andic part within 100 cm.
}
