% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_aeolic}
\alias{qual_aeolic}
\title{Aeolic qualifier (ae): aeolic (wind-sorted) material <= 100 cm.}
\usage{
qual_aeolic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Aeolic qualifier (ae): aeolic (wind-sorted) material <= 100 cm.
}
\keyword{internal}
