% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_albic}
\alias{qual_albic}
\title{Albic qualifier (ab): albic horizon <= 100 cm.}
\usage{
qual_albic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Albic qualifier (ab): albic horizon <= 100 cm.
}
