% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_alcalic}
\alias{qual_alcalic}
\title{Alcalic supplementary qualifier (ac): pH (H2O) >= 9.0
WRB 2022 Ch 5: "Strongly alkaline reaction (pH H2O >= 9 in any
layer within 100 cm of the soil surface)."}
\usage{
qual_alcalic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Alcalic supplementary qualifier (ac): pH (H2O) >= 9.0
WRB 2022 Ch 5: "Strongly alkaline reaction (pH H2O >= 9 in any
layer within 100 cm of the soil surface)."
}
