% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_alic}
\alias{qual_alic}
\title{Alic qualifier (al): argic + high CEC + high Al saturation.}
\usage{
qual_alic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Alic qualifier (al): argic + high CEC + high Al saturation.
}
