% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091b.R
\name{qual_aluandic}
\alias{qual_aluandic}
\title{Aluandic qualifier (aa): andic properties + Al-dominant active
component (Al / (Al + 0.5 Si) >= 0.5 in mass).}
\usage{
qual_aluandic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Aluandic qualifier (aa): andic properties + Al-dominant active
component (Al / (Al + 0.5 Si) >= 0.5 in mass).
}
