% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_andic}
\alias{qual_andic}
\title{Andic qualifier (an): andic OR vitric properties combined >= 30 cm.
v0.9 simplification: passes if andic_properties or vitric_properties
passes within 100 cm.}
\usage{
qual_andic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Andic qualifier (an): andic OR vitric properties combined >= 30 cm.
v0.9 simplification: passes if andic_properties or vitric_properties
passes within 100 cm.
}
