% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_anofluvic}
\alias{qual_anofluvic}
\title{Anofluvic qualifier (af): fluvic material only at depth >= 50 cm}
\usage{
qual_anofluvic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
WRB 2022 Ch 5 (Fluvisols): "Fluvic material starting >= 50 cm
from the soil surface." Depth modifier of
\code{\link{fluvic_material}}.
}
