% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_anthraquic}
\alias{qual_anthraquic}
\title{Anthraquic qualifier (aq): anthraquic horizon (puddled-rice surface).}
\usage{
qual_anthraquic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Anthraquic qualifier (aq): anthraquic horizon (puddled-rice surface).
}
