% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_anthric}
\alias{qual_anthric}
\title{Anthric qualifier (ak): anthric properties.}
\usage{
qual_anthric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Anthric qualifier (ak): anthric properties.
}
