% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_anthromollic}
\alias{qual_anthromollic}
\title{Anthromollic qualifier (am): anthric horizon overlying spodic}
\usage{
qual_anthromollic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
WRB 2022 Ch 5 (Podzols): "Having an anthric (irrigation /
Plaggic-like) surface horizon directly over spodic / albic /
diagnostic horizon." Combines \code{\link{anthric_horizons}} +
overlying-spodic check.
}
