% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_arenic}
\alias{qual_arenic}
\title{Arenic qualifier (ar): texture sand or loamy sand >= 30 cm in <= 100 cm.}
\usage{
qual_arenic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Arenic qualifier (ar): texture sand or loamy sand >= 30 cm in <= 100 cm.
}
