% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_arenicolic}
\alias{qual_arenicolic}
\title{Arenicolic supplementary qualifier (an): faunal sand burrows
Tier-3: requires bioturbation / burrow-density site field.}
\usage{
qual_arenicolic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Arenicolic supplementary qualifier (an): faunal sand burrows
Tier-3: requires bioturbation / burrow-density site field.
}
