% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v093b.R
\name{qual_aric}
\alias{qual_aric}
\title{Aric qualifier (ar): mineral surface horizon homogenised by
ploughing -- designation pattern \code{Ap}, \code{Apk},
\code{Apc}, etc., starting within the upper 30 cm.}
\usage{
qual_aric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Aric qualifier (ar): mineral surface horizon homogenised by
ploughing -- designation pattern \code{Ap}, \code{Apk},
\code{Apc}, etc., starting within the upper 30 cm.
}
