% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_bathyspodic}
\alias{qual_bathyspodic}
\title{Bathyspodic supplementary qualifier (bs): spodic at 100-200 cm depth}
\usage{
qual_bathyspodic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Bathyspodic supplementary qualifier (bs): spodic at 100-200 cm depth
}
