% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_biocrustic}
\alias{qual_biocrustic}
\title{Biocrustic supplementary qualifier (bk): biological soil crust
Tier-3: requires surface-crust morphology field.}
\usage{
qual_biocrustic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Biocrustic supplementary qualifier (bk): biological soil crust
Tier-3: requires surface-crust morphology field.
}
