% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_bryic}
\alias{qual_bryic}
\title{Bryic supplementary qualifier (by): bryophyte cover
Tier-3: requires vegetation-cover composition field.}
\usage{
qual_bryic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Bryic supplementary qualifier (by): bryophyte cover
Tier-3: requires vegetation-cover composition field.
}
