% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_calcaric}
\alias{qual_calcaric}
\title{Calcaric qualifier (cl): calcaric material >= 25 cm in upper 100 cm.}
\usage{
qual_calcaric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Calcaric qualifier (cl): calcaric material >= 25 cm in upper 100 cm.
}
