% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_calcic}
\alias{qual_calcic}
\title{Calcic qualifier (cc): calcic horizon <= 100 cm.}
\usage{
qual_calcic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Calcic qualifier (cc): calcic horizon <= 100 cm.
}
