% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_cambic}
\alias{qual_cambic}
\title{Cambic qualifier (cm): cambic horizon <= 50 cm.}
\usage{
qual_cambic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Cambic qualifier (cm): cambic horizon <= 50 cm.
}
