% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_carbonatic}
\alias{qual_carbonatic}
\title{Carbonatic supplementary qualifier (cn): >= 50\% carbonates}
\usage{
qual_carbonatic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Carbonatic supplementary qualifier (cn): >= 50\% carbonates
}
