% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_carbonic}
\alias{qual_carbonic}
\title{Carbonic supplementary qualifier (cb): high SOC content (>= 6\%)}
\usage{
qual_carbonic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Carbonic supplementary qualifier (cb): high SOC content (>= 6\%)
}
