% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091b.R
\name{qual_chernic}
\alias{qual_chernic}
\title{Chernic qualifier (ch): chernic horizon (intensely worm-mixed mollic-like)
within 100 cm.}
\usage{
qual_chernic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Chernic qualifier (ch): chernic horizon (intensely worm-mixed mollic-like)
within 100 cm.
}
