% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_chromic}
\alias{qual_chromic}
\title{Chromic qualifier (cr): hue redder than 7.5YR + chroma > 4 (in upper
subsoil 25-150 cm).}
\usage{
qual_chromic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Chromic qualifier (cr): hue redder than 7.5YR + chroma > 4 (in upper
subsoil 25-150 cm).
}
