% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_claric}
\alias{qual_claric}
\title{Claric qualifier (cl): claric (light-coloured) material <= 100 cm.}
\usage{
qual_claric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Claric qualifier (cl): claric (light-coloured) material <= 100 cm.
}
\keyword{internal}
