% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_columnic}
\alias{qual_columnic}
\title{Columnic supplementary qualifier (cm): columnar / prismatic structure
WRB 2022 Ch 5: "Columnar or strong prismatic structure
(associated with natric horizons)."}
\usage{
qual_columnic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Columnic supplementary qualifier (cm): columnar / prismatic structure
WRB 2022 Ch 5: "Columnar or strong prismatic structure
(associated with natric horizons)."
}
