% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_cryic}
\alias{qual_cryic}
\title{Cryic qualifier (cy): cryic horizon <= 100 cm.}
\usage{
qual_cryic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Cryic qualifier (cy): cryic horizon <= 100 cm.
}
