% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091c.R
\name{qual_densic}
\alias{qual_densic}
\title{Densic qualifier (dn): bulk density >= 1.8 g/cm3 in some root-
restricting layer within 100 cm.}
\usage{
qual_densic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Densic qualifier (dn): bulk density >= 1.8 g/cm3 in some root-
restricting layer within 100 cm.
}
