% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_differentic}
\alias{qual_differentic}
\title{Differentic supplementary qualifier (df): contrasting layers
WRB 2022 Ch 5: "Strong differences (texture, mineralogy, color)
between adjacent layers without abrupt textural transition (mild
clay-increase 1.2-1.4x ratio)."}
\usage{
qual_differentic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Differentic supplementary qualifier (df): contrasting layers
WRB 2022 Ch 5: "Strong differences (texture, mineralogy, color)
between adjacent layers without abrupt textural transition (mild
clay-increase 1.2-1.4x ratio)."
}
