% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_dolomitic}
\alias{qual_dolomitic}
\title{Dolomitic qualifier (do): dolomitic material in upper 100 cm.}
\usage{
qual_dolomitic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Dolomitic qualifier (do): dolomitic material in upper 100 cm.
}
