% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_dorsic}
\alias{qual_dorsic}
\title{Dorsic supplementary qualifier (do): dorsal-ridge microrelief}
\usage{
qual_dorsic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Dorsic supplementary qualifier (do): dorsal-ridge microrelief
}
