% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_duric}
\alias{qual_duric}
\title{Duric qualifier (du): duric horizon <= 100 cm.}
\usage{
qual_duric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Duric qualifier (du): duric horizon <= 100 cm.
}
