% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_ekranic}
\alias{qual_ekranic}
\title{Ekranic qualifier (ek): impervious cover (asphalt, concrete) starting
within 5 cm of the surface. v0.9.1: technic_hard_material with top
depth <= 5 cm.}
\usage{
qual_ekranic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Ekranic qualifier (ek): impervious cover (asphalt, concrete) starting
within 5 cm of the surface. v0.9.1: technic_hard_material with top
depth <= 5 cm.
}
