% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_endoabruptic}
\alias{qual_endoabruptic}
\title{Endoabruptic supplementary qualifier (eea): abrupt textural change deep}
\usage{
qual_endoabruptic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Endoabruptic supplementary qualifier (eea): abrupt textural change deep
}
