% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_endocalcic}
\alias{qual_endocalcic}
\title{Endocalcic supplementary qualifier: calcic horizon at depth >= 50 cm}
\usage{
qual_endocalcic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Endocalcic supplementary qualifier: calcic horizon at depth >= 50 cm
}
