% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_endoduric}
\alias{qual_endoduric}
\title{Endoduric supplementary qualifier: duric horizon at depth >= 50 cm}
\usage{
qual_endoduric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Endoduric supplementary qualifier: duric horizon at depth >= 50 cm
}
