% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_endodystric}
\alias{qual_endodystric}
\title{Endodystric supplementary qualifier (eds): dystric only at depth}
\usage{
qual_endodystric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
WRB 2022 Ch 5: "Distric (BS < 50\%) at >= 50 cm depth."
}
