% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0933.R
\name{qual_endogleyic}
\alias{qual_endogleyic}
\title{Endogleyic qualifier (eng): gleyic conditions between 50 and 100 cm.}
\usage{
qual_endogleyic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Endogleyic qualifier (eng): gleyic conditions between 50 and 100 cm.
}
