% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_endogypsic}
\alias{qual_endogypsic}
\title{Endogypsic supplementary qualifier: gypsic horizon at depth >= 50 cm}
\usage{
qual_endogypsic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Endogypsic supplementary qualifier: gypsic horizon at depth >= 50 cm
}
