% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_endoleptic}
\alias{qual_endoleptic}
\title{Endoleptic supplementary qualifier (lle): rock contact 50-100 cm}
\usage{
qual_endoleptic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Endoleptic supplementary qualifier (lle): rock contact 50-100 cm
}
