% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0933.R
\name{qual_endostagnic}
\alias{qual_endostagnic}
\title{Endostagnic qualifier (ens): stagnic conditions between 50 and 100 cm.}
\usage{
qual_endostagnic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Endostagnic qualifier (ens): stagnic conditions between 50 and 100 cm.
}
