% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_endothionic}
\alias{qual_endothionic}
\title{Endothionic supplementary qualifier (etn): thionic at depth >= 50 cm}
\usage{
qual_endothionic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Endothionic supplementary qualifier (etn): thionic at depth >= 50 cm
}
