% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_endothyric}
\alias{qual_endothyric}
\title{Endothyric supplementary qualifier (etc): thyric only at depth >= 50}
\usage{
qual_endothyric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Endothyric supplementary qualifier (etc): thyric only at depth >= 50
}
