% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_entic}
\alias{qual_entic}
\title{Entic qualifier (et): albic horizon AND NOT spodic}
\usage{
qual_entic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
WRB 2022 Ch 5 (Podzols): "Having an albic horizon (>= 1 cm thick)
starting <= 50 cm AND NOT meeting the criteria for a spodic
horizon." Compose: albic AND NOT spodic.
}
