% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_epic}
\alias{qual_epic}
\title{Epic supplementary qualifier (ep): generic "in shallow horizon"}
\usage{
qual_epic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Epic supplementary qualifier (ep): generic "in shallow horizon"
}
