% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_epidystric}
\alias{qual_epidystric}
\title{Epidystric supplementary qualifier (epd): dystric only in the upper part}
\usage{
qual_epidystric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
WRB 2022 Ch 5: exchangeable Al > bases in 20-50 cm (Dystric restricted to
the upper part).
}
\keyword{internal}
