% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_epieutric}
\alias{qual_epieutric}
\title{Epieutric supplementary qualifier (eee): eutric only in upper 50 cm}
\usage{
qual_epieutric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Epieutric supplementary qualifier (eee): eutric only in upper 50 cm
}
