% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_escalic}
\alias{qual_escalic}
\title{Escalic supplementary qualifier (es): terraced / stepped morphology}
\usage{
qual_escalic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Escalic supplementary qualifier (es): terraced / stepped morphology
}
