% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091b.R
\name{qual_eutrosilic}
\alias{qual_eutrosilic}
\title{Eutrosilic qualifier (es): silandic + base saturation >= 50\% in some
layer of the silandic part within 100 cm.}
\usage{
qual_eutrosilic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Eutrosilic qualifier (es): silandic + base saturation >= 50\% in some
layer of the silandic part within 100 cm.
}
