% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_evapocrustic}
\alias{qual_evapocrustic}
\title{Evapocrustic supplementary qualifier (ev): evaporite surface crust}
\usage{
qual_evapocrustic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Evapocrustic supplementary qualifier (ev): evaporite surface crust
}
