% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_ferralic}
\alias{qual_ferralic}
\title{Ferralic qualifier (fl): ferralic horizon <= 150 cm.}
\usage{
qual_ferralic(pedon)
}
\description{
Ferralic qualifier (fl): ferralic horizon <= 150 cm.
}
