% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_ferric}
\alias{qual_ferric}
\title{Ferric qualifier (fr): ferric horizon <= 100 cm.}
\usage{
qual_ferric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Ferric qualifier (fr): ferric horizon <= 100 cm.
}
