% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_fluvic}
\alias{qual_fluvic}
\title{Fluvic qualifier (fv): fluvic material >= 25 cm thick starting <= 75 cm.}
\usage{
qual_fluvic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Fluvic qualifier (fv): fluvic material >= 25 cm thick starting <= 75 cm.
}
