% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_fractic}
\alias{qual_fractic}
\title{Fractic qualifier (fc): fractures (cracks) within 100 cm}
\usage{
qual_fractic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
WRB 2022 Ch 5 (Durisols / Gypsisols / Calcisols): "Showing
fractures within 100 cm of the soil surface" (a duripan, gypsic,
or calcic horizon that has cracked / fractured).
}
\details{
Implementation: positive \code{cracks_width_cm} or
\code{cracks_depth_cm} on any layer with top <= 100 cm.
}
