% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_fragic}
\alias{qual_fragic}
\title{Fragic qualifier (fg): fragic horizon <= 100 cm.}
\usage{
qual_fragic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Fragic qualifier (fg): fragic horizon <= 100 cm.
}
\keyword{internal}
