% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_gelic}
\alias{qual_gelic}
\title{Gelic supplementary qualifier (gl): permafrost or strong frost activity
WRB 2022 Ch 5: "Permafrost within 200 cm of the soil surface OR
gelic materials." Modifier of cryic_conditions for non-Cryosols.}
\usage{
qual_gelic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Gelic supplementary qualifier (gl): permafrost or strong frost activity
WRB 2022 Ch 5: "Permafrost within 200 cm of the soil surface OR
gelic materials." Modifier of cryic_conditions for non-Cryosols.
}
