% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_gelistagnic}
\alias{qual_gelistagnic}
\title{Gelistagnic supplementary qualifier (gst): stagnic in cold conditions
WRB 2022 Ch 5: "Stagnic features (perched water) in cryic regime."
Compose: stagnic_pattern + cryic_conditions.}
\usage{
qual_gelistagnic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Gelistagnic supplementary qualifier (gst): stagnic in cold conditions
WRB 2022 Ch 5: "Stagnic features (perched water) in cryic regime."
Compose: stagnic_pattern + cryic_conditions.
}
