% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_gilgaic}
\alias{qual_gilgaic}
\title{Gilgaic supplementary qualifier (gi): gilgai microrelief
WRB 2022 Ch 5: "Gilgai microrelief (associated with vertic
shrinking/swelling soils)." Site-level field detection.}
\usage{
qual_gilgaic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Gilgaic supplementary qualifier (gi): gilgai microrelief
WRB 2022 Ch 5: "Gilgai microrelief (associated with vertic
shrinking/swelling soils)." Site-level field detection.
}
