% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_gleyic}
\alias{qual_gleyic}
\title{Gleyic qualifier (gl): gleyic properties throughout a layer >= 25 cm
starting <= 75 cm + reducing conditions.}
\usage{
qual_gleyic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Gleyic qualifier (gl): gleyic properties throughout a layer >= 25 cm
starting <= 75 cm + reducing conditions.
}
