% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091de.R
\name{qual_glossic}
\alias{qual_glossic}
\title{Glossic qualifier (gs): mollic horizon penetrated by albeluvic
tongues (glossae). Diagnostic of Glossic Chernozems / Phaeozems on
the steppe / forest-steppe transition.}
\usage{
qual_glossic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Glossic qualifier (gs): mollic horizon penetrated by albeluvic
tongues (glossae). Diagnostic of Glossic Chernozems / Phaeozems on
the steppe / forest-steppe transition.
}
