% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091b.R
\name{qual_grumic}
\alias{qual_grumic}
\title{Grumic qualifier (gr): strong fine granular surface horizon
(self-mulching Vertisol).}
\usage{
qual_grumic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Grumic qualifier (gr): strong fine granular surface horizon
(self-mulching Vertisol).
}
