% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_gypsic}
\alias{qual_gypsic}
\title{Gypsic qualifier (gy): gypsic horizon <= 100 cm.}
\usage{
qual_gypsic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Gypsic qualifier (gy): gypsic horizon <= 100 cm.
}
