% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_gypsiric}
\alias{qual_gypsiric}
\title{Gypsiric qualifier (gc): gypsiric material >= 25 cm in upper 100 cm.}
\usage{
qual_gypsiric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Gypsiric qualifier (gc): gypsiric material >= 25 cm in upper 100 cm.
}
