% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_histic}
\alias{qual_histic}
\title{Histic qualifier (hi): histic horizon at or near the surface.}
\usage{
qual_histic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Histic qualifier (hi): histic horizon at or near the surface.
}
