% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_hortic}
\alias{qual_hortic}
\title{Hortic qualifier (ht): hortic horizon (long-cultivated dark surface).}
\usage{
qual_hortic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Hortic qualifier (ht): hortic horizon (long-cultivated dark surface).
}
