% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_hydragric}
\alias{qual_hydragric}
\title{Hydragric qualifier (hg): hydragric horizon (puddled-rice subsurface).}
\usage{
qual_hydragric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Hydragric qualifier (hg): hydragric horizon (puddled-rice subsurface).
}
