% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_hydrophobic}
\alias{qual_hydrophobic}
\title{Hydrophobic supplementary qualifier (hf): water-repellent surface}
\usage{
qual_hydrophobic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
WRB 2022 Ch 5: "Surface horizon (0-5 cm) with hydrophobic
character measurable as MED (Molarity of an Ethanol Droplet) >= 1
or WDPT (Water Drop Penetration Time) >= 60 s."
Implementation: textual flag in \code{vesicular_pores} (BDsolos:
"hidrofóbico", "water repellent") OR a \code{water_repellence}
field if the loader supplies it.
}
