% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091c.R
\name{qual_hyperalic}
\alias{qual_hyperalic}
\title{Hyperalic qualifier (yl): argic horizon with Al saturation >= 50\% in
some layer of the argic part within 100 cm. Stronger version of Alic.}
\usage{
qual_hyperalic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Hyperalic qualifier (yl): argic horizon with Al saturation >= 50\% in
some layer of the argic part within 100 cm. Stronger version of Alic.
}
