% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_hyperartefactic}
\alias{qual_hyperartefactic}
\title{Hyperartefactic qualifier (yr): >= 80\% artefacts (any type) in the
upper 100 cm.}
\usage{
qual_hyperartefactic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Hyperartefactic qualifier (yr): >= 80\% artefacts (any type) in the
upper 100 cm.
}
