% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v092a.R
\name{qual_hypercalcic}
\alias{qual_hypercalcic}
\title{Hypercalcic qualifier (yc): calcic horizon AND CaCO3 >= 50\% in some
calcic layer.}
\usage{
qual_hypercalcic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Hypercalcic qualifier (yc): calcic horizon AND CaCO3 >= 50\% in some
calcic layer.
}
