% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v092a.R
\name{qual_hypergypsic}
\alias{qual_hypergypsic}
\title{Hypergypsic qualifier (yg): gypsic horizon AND gypsum >= 60\% in
some gypsic layer.}
\usage{
qual_hypergypsic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Hypergypsic qualifier (yg): gypsic horizon AND gypsum >= 60\% in
some gypsic layer.
}
