% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_hyperorganic}
\alias{qual_hyperorganic}
\title{Hyperorganic supplementary qualifier (hyo): SOC >= 18\% (peat-like)
WRB 2022 Ch 5: "Containing organic carbon >= 18\% by mass in any
layer >= 10 cm thick." A stronger version of `Carbonic`.}
\usage{
qual_hyperorganic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Hyperorganic supplementary qualifier (hyo): SOC >= 18\% (peat-like)
WRB 2022 Ch 5: "Containing organic carbon >= 18\% by mass in any
layer >= 10 cm thick." A stronger version of `Carbonic`.
}
