% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v092a.R
\name{qual_hypersalic}
\alias{qual_hypersalic}
\title{Hypersalic qualifier (yz): EC (1:5 H2O extract) >= 30 dS/m in some
layer within the upper 100 cm. Stronger than the Salic horizon
(default >= 15 dS/m).}
\usage{
qual_hypersalic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Hypersalic qualifier (yz): EC (1:5 H2O extract) >= 30 dS/m in some
layer within the upper 100 cm. Stronger than the Salic horizon
(default >= 15 dS/m).
}
