% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_hyperskeletic}
\alias{qual_hyperskeletic}
\title{Hyperskeletic qualifier (hk): coarse fragments >= 90\% throughout the
upper 100 cm.}
\usage{
qual_hyperskeletic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Hyperskeletic qualifier (hk): coarse fragments >= 90\% throughout the
upper 100 cm.
}
