% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v092a.R
\name{qual_hypersodic}
\alias{qual_hypersodic}
\title{Hypersodic qualifier (yo): ESP >= 50\% in some layer within 100 cm.
Stronger than Sodic (default ESP >= 6\%).}
\usage{
qual_hypersodic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Hypersodic qualifier (yo): ESP >= 50\% in some layer within 100 cm.
Stronger than Sodic (default ESP >= 6\%).
}
