% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v092a.R
\name{qual_hyposalic}
\alias{qual_hyposalic}
\title{Hyposalic qualifier (jz): EC (1:5 H2O extract) >= 4 dS/m AND < 15
dS/m in some layer within the upper 100 cm. Used for soils too
weak to qualify as Solonchak but still carrying a salinity tag.}
\usage{
qual_hyposalic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Hyposalic qualifier (jz): EC (1:5 H2O extract) >= 4 dS/m AND < 15
dS/m in some layer within the upper 100 cm. Used for soils too
weak to qualify as Solonchak but still carrying a salinity tag.
}
