% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v092a.R
\name{qual_hyposodic}
\alias{qual_hyposodic}
\title{Hyposodic qualifier (jo): ESP >= 6\% AND < 15\% in some layer within
100 cm. Marginal sodicity tag.}
\usage{
qual_hyposodic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Hyposodic qualifier (jo): ESP >= 6\% AND < 15\% in some layer within
100 cm. Marginal sodicity tag.
}
