% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_immissic}
\alias{qual_immissic}
\title{Immissic supplementary qualifier (im): atmospheric immission}
\usage{
qual_immissic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Immissic supplementary qualifier (im): atmospheric immission
}
