% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_inclinic}
\alias{qual_inclinic}
\title{Inclinic supplementary qualifier (in): tilted / inclined position
WRB 2022 Ch 5: site has a slope >= 10\% (relevo declivoso).
Implementation: site$slope_pct (when populated) >= 10 OR
parent_material / forma_relevo flagging steep terrain.}
\usage{
qual_inclinic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Inclinic supplementary qualifier (in): tilted / inclined position
WRB 2022 Ch 5: site has a slope >= 10\% (relevo declivoso).
Implementation: site$slope_pct (when populated) >= 10 OR
parent_material / forma_relevo flagging steep terrain.
}
