% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_irragric}
\alias{qual_irragric}
\title{Irragric qualifier (ir): irragric horizon (irrigation-deposited surface).}
\usage{
qual_irragric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Irragric qualifier (ir): irragric horizon (irrigation-deposited surface).
}
