% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_isopteric}
\alias{qual_isopteric}
\title{Isopteric supplementary qualifier (ip): termite / ant biogenesis}
\usage{
qual_isopteric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Isopteric supplementary qualifier (ip): termite / ant biogenesis
}
