% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_kalaic}
\alias{qual_kalaic}
\title{Kalaic supplementary qualifier (ka): dry-season puffed surface layer}
\usage{
qual_kalaic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Kalaic supplementary qualifier (ka): dry-season puffed surface layer
}
