% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v093b.R
\name{qual_lamellic}
\alias{qual_lamellic}
\title{Lamellic qualifier (ll): thin (\eqn{<} 5 cm) clay-enriched
lamellae, typical of sandy Luvisols / Alisols / Acrisols.
v0.9.3.B proxy: designation pattern \code{lamell} / \code{E&Bt} /
\code{&Bt} / \code{Bt(t)?\\d?lam} in any subsurface layer.}
\usage{
qual_lamellic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Lamellic qualifier (ll): thin (\eqn{<} 5 cm) clay-enriched
lamellae, typical of sandy Luvisols / Alisols / Acrisols.
v0.9.3.B proxy: designation pattern \code{lamell} / \code{E&Bt} /
\code{&Bt} / \code{Bt(t)?\\d?lam} in any subsurface layer.
}
