% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_lapiadic}
\alias{qual_lapiadic}
\title{Lapiadic supplementary qualifier (lp): karren / lapies bedrock features}
\usage{
qual_lapiadic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Lapiadic supplementary qualifier (lp): karren / lapies bedrock features
}
