% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_laxic}
\alias{qual_laxic}
\title{Laxic supplementary qualifier (lx): loose / non-cohesive surface
WRB 2022 Ch 5: "Surface horizon with loose dry consistence and
single-grain or massive structure."}
\usage{
qual_laxic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Laxic supplementary qualifier (lx): loose / non-cohesive surface
WRB 2022 Ch 5: "Surface horizon with loose dry consistence and
single-grain or massive structure."
}
