% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_leptic}
\alias{qual_leptic}
\title{Leptic qualifier (le): continuous rock <= 100 cm.}
\usage{
qual_leptic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Leptic qualifier (le): continuous rock <= 100 cm.
}
