% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_limnic}
\alias{qual_limnic}
\title{Limnic qualifier (lm): limnic material (lacustrine / marine subaquatic
deposits) anywhere in the profile.}
\usage{
qual_limnic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Limnic qualifier (lm): limnic material (lacustrine / marine subaquatic
deposits) anywhere in the profile.
}
