% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_limonic}
\alias{qual_limonic}
\title{Limonic qualifier (lm): limonic (bog-iron) horizon <= 100 cm.}
\usage{
qual_limonic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Limonic qualifier (lm): limonic (bog-iron) horizon <= 100 cm.
}
\keyword{internal}
