% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_linic}
\alias{qual_linic}
\title{Linic qualifier (li): continuous artificial geomembrane within 100 cm.
v0.9.1 proxy: designation pattern (\code{linic|geomemb|liner}).}
\usage{
qual_linic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Linic qualifier (li): continuous artificial geomembrane within 100 cm.
v0.9.1 proxy: designation pattern (\code{linic|geomemb|liner}).
}
