% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_lithic}
\alias{qual_lithic}
\title{Lithic qualifier (lt): continuous rock starting within 10 cm. Tighter
depth gate than Leptic (which is <= 100 cm) and Nudilithic (== 0 cm).}
\usage{
qual_lithic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Lithic qualifier (lt): continuous rock starting within 10 cm. Tighter
depth gate than Leptic (which is <= 100 cm) and Nudilithic (== 0 cm).
}
