% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_litholinic}
\alias{qual_litholinic}
\title{Litholinic supplementary qualifier (ll): stratified soil on rock}
\usage{
qual_litholinic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Litholinic supplementary qualifier (ll): stratified soil on rock
}
