% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_lixic}
\alias{qual_lixic}
\title{Lixic qualifier (lx): argic + low CEC, low Al.}
\usage{
qual_lixic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Lixic qualifier (lx): argic + low CEC, low Al.
}
