% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_loamic}
\alias{qual_loamic}
\title{Loamic qualifier (lo): loam-class texture >= 30 cm in the upper 100 cm.}
\usage{
qual_loamic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Loamic qualifier (lo): loam-class texture >= 30 cm in the upper 100 cm.
}
