% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_luvic}
\alias{qual_luvic}
\title{Luvic qualifier (lv): argic + high CEC, low Al saturation.}
\usage{
qual_luvic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Luvic qualifier (lv): argic + high CEC, low Al saturation.
}
