% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_magnesic}
\alias{qual_magnesic}
\title{Magnesic qualifier (mg): exchangeable Ca/Mg < 1 in upper 100 cm.}
\usage{
qual_magnesic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Magnesic qualifier (mg): exchangeable Ca/Mg < 1 in upper 100 cm.
}
