% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091b.R
\name{qual_mazic}
\alias{qual_mazic}
\title{Mazic qualifier (mz): structureless / massive surface horizon
(Vertisol). Diagnostic of slaked, crusted Vertisol surfaces.}
\usage{
qual_mazic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Mazic qualifier (mz): structureless / massive surface horizon
(Vertisol). Diagnostic of slaked, crusted Vertisol surfaces.
}
