% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_mineralic}
\alias{qual_mineralic}
\title{Mineralic supplementary qualifier (mn): predominantly mineral
WRB 2022 Ch 5: "Predominantly mineral material in upper 100 cm
(oc_pct < 12\% averaged over depth)."}
\usage{
qual_mineralic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Mineralic supplementary qualifier (mn): predominantly mineral
WRB 2022 Ch 5: "Predominantly mineral material in upper 100 cm
(oc_pct < 12\% averaged over depth)."
}
