% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_mochipic}
\alias{qual_mochipic}
\title{Mochipic supplementary qualifier (mp): mottled mochi-like pattern}
\usage{
qual_mochipic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Mochipic supplementary qualifier (mp): mottled mochi-like pattern
}
