% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_mollic}
\alias{qual_mollic}
\title{Mollic qualifier (mo): mollic horizon.}
\usage{
qual_mollic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Mollic qualifier (mo): mollic horizon.
}
