% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_mulmic}
\alias{qual_mulmic}
\title{Mulmic qualifier (ml): mulmic material in upper 100 cm.}
\usage{
qual_mulmic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Mulmic qualifier (ml): mulmic material in upper 100 cm.
}
