% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_naramic}
\alias{qual_naramic}
\title{Naramic supplementary qualifier (na): salt-crust morphology}
\usage{
qual_naramic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Naramic supplementary qualifier (na): salt-crust morphology
}
