% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_natric}
\alias{qual_natric}
\title{Natric qualifier (na): natric horizon <= 100 cm.}
\usage{
qual_natric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Natric qualifier (na): natric horizon <= 100 cm.
}
