% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_nechic}
\alias{qual_nechic}
\title{Nechic supplementary qualifier (ne): aeolian / loess deposit pattern}
\usage{
qual_nechic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Nechic supplementary qualifier (ne): aeolian / loess deposit pattern
}
