% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_nitic}
\alias{qual_nitic}
\title{Nitic qualifier (ni): nitic horizon <= 100 cm.}
\usage{
qual_nitic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Nitic qualifier (ni): nitic horizon <= 100 cm.
}
