% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_nudilithic}
\alias{qual_nudilithic}
\title{Nudilithic qualifier (nt): continuous rock at the soil surface (top_cm == 0).}
\usage{
qual_nudilithic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Nudilithic qualifier (nt): continuous rock at the soil surface (top_cm == 0).
}
