% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_organotechnic}
\alias{qual_organotechnic}
\title{Organotechnic qualifier (ot): organotechnic material in upper 100 cm.}
\usage{
qual_organotechnic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Organotechnic qualifier (ot): organotechnic material in upper 100 cm.
}
