% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_ornithic}
\alias{qual_ornithic}
\title{Ornithic qualifier (oc): ornithogenic material (bird-influenced topsoil)
in the upper 50 cm.}
\usage{
qual_ornithic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Ornithic qualifier (oc): ornithogenic material (bird-influenced topsoil)
in the upper 50 cm.
}
