% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091c.R
\name{qual_ortsteinic}
\alias{qual_ortsteinic}
\title{Ortsteinic qualifier (os): cemented spodic horizon. v0.9.1:
spodic horizon + cementation_class strongly OR indurated.}
\usage{
qual_ortsteinic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Ortsteinic qualifier (os): cemented spodic horizon. v0.9.1:
spodic horizon + cementation_class strongly OR indurated.
}
