% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_oxygleyic}
\alias{qual_oxygleyic}
\title{Oxygleyic qualifier (og): gleyic regime with predominant oxidation}
\usage{
qual_oxygleyic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
WRB 2022 Ch 5 (Gleysols): "Gleyic properties dominated by oxidation
(redox concentrations >> reductive depletions)." Heuristic: gleyic
fires AND redoximorphic_features_pct >= 10 in upper 50 cm.
}
