% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091b.R
\name{qual_pachic}
\alias{qual_pachic}
\title{Pachic qualifier (pc): mollic OR umbric horizon >= 50 cm thick.}
\usage{
qual_pachic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Pachic qualifier (pc): mollic OR umbric horizon >= 50 cm thick.
}
