% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_panpaic}
\alias{qual_panpaic}
\title{Panpaic qualifier (pp): a panpan (cemented/indurated pan) <= 100 cm.}
\usage{
qual_panpaic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Panpaic qualifier (pp): a panpan (cemented/indurated pan) <= 100 cm.
}
\keyword{internal}
