% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_pantofluvic}
\alias{qual_pantofluvic}
\title{Pantofluvic qualifier (pf): fluvic material throughout 0-100 cm}
\usage{
qual_pantofluvic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
WRB 2022 Ch 5 (Fluvisols): "Fluvic material continuously from the
soil surface to >= 100 cm depth."
}
