% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091b.R
\name{qual_pellic}
\alias{qual_pellic}
\title{Pellic qualifier (pe): in the upper 30 cm, Munsell value <= 4 moist
AND chroma <= 2 moist. Diagnostic of "black" (dark) Vertisols.}
\usage{
qual_pellic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Pellic qualifier (pe): in the upper 30 cm, Munsell value <= 4 moist
AND chroma <= 2 moist. Diagnostic of "black" (dark) Vertisols.
}
