% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_pelocrustic}
\alias{qual_pelocrustic}
\title{Pelocrustic supplementary qualifier (pc): clayey surface crust}
\usage{
qual_pelocrustic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Pelocrustic supplementary qualifier (pc): clayey surface crust
}
