% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_petric}
\alias{qual_petric}
\title{Petric qualifier (pt): any petro-cemented horizon (petrocalcic /
petroduric / petrogypsic / petroplinthic) within 100 cm.}
\usage{
qual_petric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Petric qualifier (pt): any petro-cemented horizon (petrocalcic /
petroduric / petrogypsic / petroplinthic) within 100 cm.
}
