% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_petrocalcic}
\alias{qual_petrocalcic}
\title{Petrocalcic qualifier (pc): petrocalcic horizon <= 100 cm.}
\usage{
qual_petrocalcic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Petrocalcic qualifier (pc): petrocalcic horizon <= 100 cm.
}
