% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_petroduric}
\alias{qual_petroduric}
\title{Petroduric qualifier (pd): petroduric horizon <= 100 cm.}
\usage{
qual_petroduric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Petroduric qualifier (pd): petroduric horizon <= 100 cm.
}
