% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_petrogypsic}
\alias{qual_petrogypsic}
\title{Petrogypsic qualifier (pg): petrogypsic horizon <= 100 cm.}
\usage{
qual_petrogypsic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Petrogypsic qualifier (pg): petrogypsic horizon <= 100 cm.
}
