% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_petroplinthic}
\alias{qual_petroplinthic}
\title{Petroplinthic qualifier (pp): petroplinthic horizon <= 100 cm.}
\usage{
qual_petroplinthic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Petroplinthic qualifier (pp): petroplinthic horizon <= 100 cm.
}
