% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091b.R
\name{qual_pisoplinthic}
\alias{qual_pisoplinthic}
\title{Pisoplinthic qualifier (px): pisoplinthic horizon within 100 cm.}
\usage{
qual_pisoplinthic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Pisoplinthic qualifier (px): pisoplinthic horizon within 100 cm.
}
