% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_plinthic}
\alias{qual_plinthic}
\title{Plinthic qualifier (pl): plinthic horizon <= 100 cm.}
\usage{
qual_plinthic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Plinthic qualifier (pl): plinthic horizon <= 100 cm.
}
