% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091c.R
\name{qual_posic}
\alias{qual_posic}
\title{Posic qualifier (po): net positive permanent charge (pH_KCl > pH_H2O)
in some layer at <= 100 cm. Diagnostic of the most weathered
Ferralsols where free Fe / Al oxides dominate the surface charge.}
\usage{
qual_posic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Posic qualifier (po): net positive permanent charge (pH_KCl > pH_H2O)
in some layer at <= 100 cm. Diagnostic of the most weathered
Ferralsols where free Fe / Al oxides dominate the surface charge.
}
