% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_pretic}
\alias{qual_pretic}
\title{Pretic qualifier (pt): pretic (pre-Columbian Amerindian dark earth) horizon.}
\usage{
qual_pretic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Pretic qualifier (pt): pretic (pre-Columbian Amerindian dark earth) horizon.
}
