% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091de.R
\name{qual_protic}
\alias{qual_protic}
\title{Protic qualifier (pr): Arenosol (or Regosol) with NO incipient
subsurface horizon -- i.e. an A-over-C profile where no cambic, no
argic, no spodic, no ferralic, no nitic horizon is present in the
upper 100 cm. v0.9.1 implements as the conjunction of the "no B
horizon" diagnostics.}
\usage{
qual_protic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Protic qualifier (pr): Arenosol (or Regosol) with NO incipient
subsurface horizon -- i.e. an A-over-C profile where no cambic, no
argic, no spodic, no ferralic, no nitic horizon is present in the
upper 100 cm. v0.9.1 implements as the conjunction of the "no B
horizon" diagnostics.
}
