% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_protoandic}
\alias{qual_protoandic}
\title{Protoandic supplementary qualifier (pan): early-stage andic
WRB 2022 Ch 5: "Andic-like properties below the strict threshold
(oxalate Al+Fe 0.4-2.0\%)."}
\usage{
qual_protoandic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Protoandic supplementary qualifier (pan): early-stage andic
WRB 2022 Ch 5: "Andic-like properties below the strict threshold
(oxalate Al+Fe 0.4-2.0\%)."
}
