% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_protoargic}
\alias{qual_protoargic}
\title{Protoargic supplementary qualifier (pra): early-stage argic
WRB 2022 Ch 5: "Clay increase 2-6 percentage points (below the
canonical argic threshold)."}
\usage{
qual_protoargic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Protoargic supplementary qualifier (pra): early-stage argic
WRB 2022 Ch 5: "Clay increase 2-6 percentage points (below the
canonical argic threshold)."
}
