% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v092a.R
\name{qual_protocalcic}
\alias{qual_protocalcic}
\title{Protocalcic qualifier (qc): protocalcic properties (incipient
carbonate accumulation) within the upper 100 cm. Wraps
\code{\link{protocalcic_properties}}.}
\usage{
qual_protocalcic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Protocalcic qualifier (qc): protocalcic properties (incipient
carbonate accumulation) within the upper 100 cm. Wraps
\code{\link{protocalcic_properties}}.
}
