% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v092a.R
\name{qual_protogypsic}
\alias{qual_protogypsic}
\title{Protogypsic qualifier (qg): protogypsic properties (incipient
gypsum accumulation) within the upper 100 cm. Wraps
\code{\link{protogypsic_properties}}.}
\usage{
qual_protogypsic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Protogypsic qualifier (qg): protogypsic properties (incipient
gypsum accumulation) within the upper 100 cm. Wraps
\code{\link{protogypsic_properties}}.
}
