% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_puffic}
\alias{qual_puffic}
\title{Puffic supplementary qualifier (pf): puffed surface}
\usage{
qual_puffic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Puffic supplementary qualifier (pf): puffed surface
}
