% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_raptic}
\alias{qual_raptic}
\title{Raptic supplementary qualifier (rp): stratification break}
\usage{
qual_raptic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Raptic supplementary qualifier (rp): stratification break
}
