% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_reductic}
\alias{qual_reductic}
\title{Reductic qualifier (rd): permanently reducing conditions caused by
anthropogenic gas / liquid emissions (typical of Technosols on
landfills). v0.9.1: reducing_conditions + Technic context.}
\usage{
qual_reductic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Reductic qualifier (rd): permanently reducing conditions caused by
anthropogenic gas / liquid emissions (typical of Technosols on
landfills). v0.9.1: reducing_conditions + Technic context.
}
