% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_reductigleyic}
\alias{qual_reductigleyic}
\title{Reductigleyic qualifier (rg): gleyic + reductive}
\usage{
qual_reductigleyic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
WRB 2022 Ch 5 (Gleysols): "Gleyic dominated by reduction
(gleyic-hue layers occupying \\>= 50\% of the upper 50 cm)."
}
