% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_rendzic}
\alias{qual_rendzic}
\title{Rendzic qualifier (rz): mollic horizon directly over calcaric material
(or limestone), shallow. Defined as Mollic + (Calcaric OR continuous
rock with carbonate parent material).}
\usage{
qual_rendzic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Rendzic qualifier (rz): mollic horizon directly over calcaric material
(or limestone), shallow. Defined as Mollic + (Calcaric OR continuous
rock with carbonate parent material).
}
