% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_retic}
\alias{qual_retic}
\title{Retic qualifier (rt): retic properties <= 100 cm.}
\usage{
qual_retic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Retic qualifier (rt): retic properties <= 100 cm.
}
