% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_rhodic}
\alias{qual_rhodic}
\title{Rhodic qualifier (ro): hue redder than 5YR + value < 4 + dry no
more than 1 unit higher than moist (in upper subsoil 25-150 cm).}
\usage{
qual_rhodic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Rhodic qualifier (ro): hue redder than 5YR + value < 4 + dry no
more than 1 unit higher than moist (in upper subsoil 25-150 cm).
}
