% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_rockic}
\alias{qual_rockic}
\title{Rockic qualifier (rk): rock-dominated organic horizon}
\usage{
qual_rockic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
WRB 2022 Ch 5 (Histosols): "Having a continuous rock or rock-like
material starting <= 25 cm from the soil surface AND >= 50\% by
volume coarse fragments in the upper 50 cm." Reuses
\code{\link{leptic_features}} (max_depth = 25) AND coarse-frag check.
}
