% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_salic}
\alias{qual_salic}
\title{Salic qualifier (sz): salic horizon <= 100 cm.}
\usage{
qual_salic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Salic qualifier (sz): salic horizon <= 100 cm.
}
