% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_saprolithic}
\alias{qual_saprolithic}
\title{Saprolithic supplementary qualifier (sp): saprolite parent material}
\usage{
qual_saprolithic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Saprolithic supplementary qualifier (sp): saprolite parent material
}
