% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_sideralic}
\alias{qual_sideralic}
\title{Sideralic qualifier (sl): sideralic properties <= 100 cm.}
\usage{
qual_sideralic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Sideralic qualifier (sl): sideralic properties <= 100 cm.
}
\keyword{internal}
