% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091b.R
\name{qual_silandic}
\alias{qual_silandic}
\title{Silandic qualifier (sn): andic properties + Si-dominant active
component (Al / (Al + 0.5 Si) < 0.5 in mass; allophane-rich).}
\usage{
qual_silandic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Silandic qualifier (sn): andic properties + Si-dominant active
component (Al / (Al + 0.5 Si) < 0.5 in mass; allophane-rich).
}
