% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_siltic}
\alias{qual_siltic}
\title{Siltic qualifier (sl): silt or silt-loam texture >= 30 cm in the upper
100 cm.}
\usage{
qual_siltic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Siltic qualifier (sl): silt or silt-loam texture >= 30 cm in the upper
100 cm.
}
