% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_solimovic}
\alias{qual_solimovic}
\title{Solimovic qualifier (sv): solimovic material (mass-movement deposits).}
\usage{
qual_solimovic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Solimovic qualifier (sv): solimovic material (mass-movement deposits).
}
