% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_someric}
\alias{qual_someric}
\title{Someric qualifier (sm): anthric epipedon over chernic / mollic}
\usage{
qual_someric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
WRB 2022 Ch 5 (Phaeozems / Chernozems / Kastanozems / Umbrisols):
"Anthric epipedon (irrigation- or Plaggic-derived) overlying a
chernic or mollic horizon." Composes anthric_horizons + mollic
(or umbric).
}
