% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_spodic}
\alias{qual_spodic}
\title{Spodic qualifier (sd): spodic horizon <= 200 cm.}
\usage{
qual_spodic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Spodic qualifier (sd): spodic horizon <= 200 cm.
}
