% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_stagnic}
\alias{qual_stagnic}
\title{Stagnic qualifier (st): stagnic properties <= 75 cm.}
\usage{
qual_stagnic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Stagnic qualifier (st): stagnic properties <= 75 cm.
}
