% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_subaquatic}
\alias{qual_subaquatic}
\title{Subaquatic qualifier (sq): permanently under water. v0.9.1:
site$drainage_class == "subaquatic" or "submerged".}
\usage{
qual_subaquatic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Subaquatic qualifier (sq): permanently under water. v0.9.1:
site$drainage_class == "subaquatic" or "submerged".
}
