% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_sulfatic}
\alias{qual_sulfatic}
\title{Sulfatic supplementary qualifier (su): high sulfate content
WRB 2022 Ch 5: "Containing >= 25\% gypsum or >= 5\% sulfate by mass."}
\usage{
qual_sulfatic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Sulfatic supplementary qualifier (su): high sulfate content
WRB 2022 Ch 5: "Containing >= 25\% gypsum or >= 5\% sulfate by mass."
}
