% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_sulfidic}
\alias{qual_sulfidic}
\title{Sulfidic qualifier (sf): hyper- OR hyposulfidic material in upper 100 cm
(the WRB Sulfidic qualifier covers either acidification class).}
\usage{
qual_sulfidic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Sulfidic qualifier (sf): hyper- OR hyposulfidic material in upper 100 cm
(the WRB Sulfidic qualifier covers either acidification class).
}
