% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_takyric}
\alias{qual_takyric}
\title{Takyric qualifier (ty): takyric properties in upper 50 cm.}
\usage{
qual_takyric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Takyric qualifier (ty): takyric properties in upper 50 cm.
}
