% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_technic}
\alias{qual_technic}
\title{Technic qualifier (tc): >= 20\% artefacts in upper 100 cm OR equivalent
geomembrane / technic-hard cover.}
\usage{
qual_technic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Technic qualifier (tc): >= 20\% artefacts in upper 100 cm OR equivalent
geomembrane / technic-hard cover.
}
