% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_tephric}
\alias{qual_tephric}
\title{Tephric qualifier (tf): tephric material >= 30 cm in upper 100 cm.}
\usage{
qual_tephric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Tephric qualifier (tf): tephric material >= 30 cm in upper 100 cm.
}
