% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_terric}
\alias{qual_terric}
\title{Terric qualifier (te): terric horizon (anthropogenic added mineral
material on top of cultivated land).}
\usage{
qual_terric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Terric qualifier (te): terric horizon (anthropogenic added mineral
material on top of cultivated land).
}
