% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_thionic}
\alias{qual_thionic}
\title{Thionic qualifier (tn): thionic horizon within 100 cm.}
\usage{
qual_thionic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Thionic qualifier (tn): thionic horizon within 100 cm.
}
