% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_thixotropic}
\alias{qual_thixotropic}
\title{Thixotropic supplementary qualifier (tx): thixotropic behavior}
\usage{
qual_thixotropic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Thixotropic supplementary qualifier (tx): thixotropic behavior
}
