% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_tidalic}
\alias{qual_tidalic}
\title{Tidalic qualifier (td): subject to tidal flooding. v0.9.1:
site$drainage_class contains "tidal".}
\usage{
qual_tidalic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Tidalic qualifier (td): subject to tidal flooding. v0.9.1:
site$drainage_class contains "tidal".
}
