% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_tsitelic}
\alias{qual_tsitelic}
\title{Tsitelic qualifier (ts): tsitelic (red, low-activity) horizon <= 100 cm.}
\usage{
qual_tsitelic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Tsitelic qualifier (ts): tsitelic (red, low-activity) horizon <= 100 cm.
}
\keyword{internal}
