% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_turbic}
\alias{qual_turbic}
\title{Turbic qualifier (tb): cryoturbation features within 100 cm. v0.9.1
proxy: cryic conditions + designation pattern (\code{turb|jj|cryot})
OR slickensides "common"/"many" in a cryic profile.}
\usage{
qual_turbic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Turbic qualifier (tb): cryoturbation features within 100 cm. v0.9.1
proxy: cryic conditions + designation pattern (\code{turb|jj|cryot})
OR slickensides "common"/"many" in a cryic profile.
}
