% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_umbric}
\alias{qual_umbric}
\title{Umbric qualifier (um): umbric horizon.}
\usage{
qual_umbric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Umbric qualifier (um): umbric horizon.
}
