% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_urbic}
\alias{qual_urbic}
\title{Urbic qualifier (ub): >= 20\% urbic artefacts (rubble, refuse) in the
upper 100 cm.}
\usage{
qual_urbic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Urbic qualifier (ub): >= 20\% urbic artefacts (rubble, refuse) in the
upper 100 cm.
}
