% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_uterquic}
\alias{qual_uterquic}
\title{Uterquic supplementary qualifier (uq): bidirectional water regime}
\usage{
qual_uterquic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Uterquic supplementary qualifier (uq): bidirectional water regime
}
