% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_vermic}
\alias{qual_vermic}
\title{Vermic qualifier (vm): >= 50\% bioturbation by worm casts / krotovinas
in the upper 100 cm. v0.9.1: \code{worm_holes_pct >= 50}.}
\usage{
qual_vermic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Vermic qualifier (vm): >= 50\% bioturbation by worm casts / krotovinas
in the upper 100 cm. v0.9.1: \code{worm_holes_pct >= 50}.
}
