% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_vertic}
\alias{qual_vertic}
\title{Vertic qualifier (vr): vertic horizon <= 100 cm.}
\usage{
qual_vertic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Vertic qualifier (vr): vertic horizon <= 100 cm.
}
