% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_vitric}
\alias{qual_vitric}
\title{Vitric qualifier (vi): vitric properties >= 30 cm within 100 cm.}
\usage{
qual_vitric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Vitric qualifier (vi): vitric properties >= 30 cm within 100 cm.
}
