% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_wapnic}
\alias{qual_wapnic}
\title{Wapnic qualifier (wp): soft, moist limnic material >= 80\% CaCO3}
\usage{
qual_wapnic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
WRB 2022 Ch 5 (Calcisols / Gleysols / Cryosols): "Having soft,
moist limnic material that contains >= 80\% by mass CaCO3
equivalent within 100 cm of the soil surface."
}
\details{
Implementation: \code{caco3_pct} >= 80 in any layer with top <= 100.
}
