% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_xanthic}
\alias{qual_xanthic}
\title{Xanthic qualifier (xa): ferralic + hue 7.5YR or yellower + value >=
4 + chroma >= 5.}
\usage{
qual_xanthic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Xanthic qualifier (xa): ferralic + hue 7.5YR or yellower + value >=
4 + chroma >= 5.
}
