% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091.R
\name{qual_yermic}
\alias{qual_yermic}
\title{Yermic qualifier (ye): yermic properties in upper 50 cm.}
\usage{
qual_yermic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Yermic qualifier (ye): yermic properties in upper 50 cm.
}
