% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-mollisols-usda.R
\name{rendoll_qualifying_usda}
\alias{rendoll_qualifying_usda}
\title{Rendolls qualifier: shallow soil over carbonate parent material.
Pass when CaCO3 >= 40\% in subsurface AND profile depth < 100 cm
to a contact.}
\usage{
rendoll_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Rendolls qualifier: shallow soil over carbonate parent material.
Pass when CaCO3 >= 40\% in subsurface AND profile depth < 100 cm
to a contact.
}
