% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-oxisols-usda.R
\name{rhodic_subgroup_usda}
\alias{rhodic_subgroup_usda}
\title{Rhodic Subgroup helper (Oxisols, Mollisols, etc.)
Pass when 50\%+ colors have hue <= 2.5YR AND value <= 3 in
B horizons 25-125 cm.}
\usage{
rhodic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Rhodic Subgroup helper (Oxisols, Mollisols, etc.)
Pass when 50\%+ colors have hue <= 2.5YR AND value <= 3 in
B horizons 25-125 cm.
}
