% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-gelisols-usda.R
\name{ruptic_subgroup_usda}
\alias{ruptic_subgroup_usda}
\title{Ruptic Subgroup helper (Histoturbels / Historthels)}
\usage{
ruptic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Pass when more than 40\% (volume) organic soil materials from
surface to 50 cm in 75\% or LESS of the pedon. v0.8 also
deferred -- requires multi-pedon data.
}
