% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-gelisols-usda.R
\name{salic_horizon_usda}
\alias{salic_horizon_usda}
\title{Salic horizon (USDA, delegates to WRB salic).}
\usage{
salic_horizon_usda(pedon, max_top_cm = 100)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{max_top_cm}{Numeric threshold or option (see Details).}
}
\description{
Salic horizon (USDA, delegates to WRB salic).
}
