% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-vertisols-usda.R
\name{salic_subgroup_usda}
\alias{salic_subgroup_usda}
\title{Salic Subgroup helper
Wraps salic_horizon_usda. Used for Salaquerts/Salitorrerts/etc.}
\usage{
salic_subgroup_usda(pedon, max_top_cm = 100)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{max_top_cm}{Numeric threshold or option (see Details).}
}
\description{
Salic Subgroup helper
Wraps salic_horizon_usda. Used for Salaquerts/Salitorrerts/etc.
}
