% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-histosols-usda.R
\name{sapric_predominant_usda}
\alias{sapric_predominant_usda}
\title{Sapric_predominant_usda: Saprists Suborder qualifier
Pass when thickness of sapric > thickness of fibric+hemic in 0-130 cm.}
\usage{
sapric_predominant_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Sapric_predominant_usda: Saprists Suborder qualifier
Pass when thickness of sapric > thickness of fibric+hemic in 0-130 cm.
}
