% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-histosols-usda.R
\name{sapric_subgroup_usda}
\alias{sapric_subgroup_usda}
\title{Sapric Subgroup helper (Sphagnofibrists)}
\usage{
sapric_subgroup_usda(pedon, max_top_cm = 130)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{max_top_cm}{Numeric threshold or option (see Details).}
}
\description{
Sapric Subgroup helper (Sphagnofibrists)
}
