% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-atributos-sibcs.R
\name{saprico}
\alias{saprico}
\title{Material organico saprico (SiBCS Cap 14)}
\usage{
saprico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
\code{\link{DiagnosticResult}}.
}
\description{
Material organico altamente decomposto: < 17\% de fibras esfregadas
OU indice de von Post H7-H10. Discrimina Organossolos Sapricos no
3o nivel categorico.
}
\references{
Embrapa (2018), SiBCS 5a ed., Cap 14 (Organossolos),
            pp 224-226.
}
