% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-histosols-usda.R
\name{smr_aridic_usda}
\alias{smr_aridic_usda}
\title{Aridic soil moisture regime (USDA)}
\usage{
smr_aridic_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Aridic soil moisture regime (USDA)
}
