% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-histosols-usda.R
\name{smr_ustic_usda}
\alias{smr_ustic_usda}
\title{Ustic soil moisture regime (USDA)}
\usage{
smr_ustic_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Ustic soil moisture regime (USDA)
}
