% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-aridisols-usda.R
\name{sodic_subgroup_usda}
\alias{sodic_subgroup_usda}
\title{Sodic Subgroup helper -- delegate to natric_horizon (USDA)}
\usage{
sodic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Sodic Subgroup helper -- delegate to natric_horizon (USDA)
}
