% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-andisols-usda.R
\name{spodic_andisol_usda}
\alias{spodic_andisol_usda}
\title{Spodic-Andisols Subgroup helper
Pass when albic horizon overlies a cambic OR spodic horizon,
OR when a spodic horizon is present in 50\%+ of the pedon.}
\usage{
spodic_andisol_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Spodic-Andisols Subgroup helper
Pass when albic horizon overlies a cambic OR spodic horizon,
OR when a spodic horizon is present in 50\%+ of the pedon.
}
