% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-orders-usda.R
\name{spodosol_usda}
\alias{spodosol_usda}
\title{Spodosols (USDA Cap 14): spodic horizon (illuvial Al/Fe/OC).}
\usage{
spodosol_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Spodosols (USDA Cap 14): spodic horizon (illuvial Al/Fe/OC).
}
