% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-histosols-usda.R
\name{str_cryic_usda}
\alias{str_cryic_usda}
\title{Cryic soil temperature regime (USDA)}
\usage{
str_cryic_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Cryic soil temperature regime (USDA)
}
