% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-spodosols-usda.R
\name{str_gelic_usda}
\alias{str_gelic_usda}
\title{Gelic soil temperature regime (USDA)}
\usage{
str_gelic_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Gelic soil temperature regime (USDA)
}
