% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-histosols-usda.R
\name{sulfic_subgroup_usda}
\alias{sulfic_subgroup_usda}
\title{Sulfic Subgroup helper (Haplowassists)
Pass when sulfidic materials within 100 cm.}
\usage{
sulfic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Sulfic Subgroup helper (Haplowassists)
Pass when sulfidic materials within 100 cm.
}
