% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-gelisols-usda.R
\name{terric_usda}
\alias{terric_usda}
\title{Terric Subgroup helper (Histels)}
\usage{
terric_usda(pedon, min_thickness_cm = 30, max_top_cm = 100)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{min_thickness_cm}{Default 30.}

\item{max_top_cm}{Default 100.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Pass when a layer of mineral soil material 30 cm or more thick
occurs within 100 cm of the soil surface (KST 13ed, p 190).
}
\references{
Soil Survey Staff (2022), KST 13ed, Ch. 9.
}
