% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-diagnostic-tests-v033.R
\name{test_cemented}
\alias{test_cemented}
\title{Test that a layer is at least moderately cemented}
\usage{
test_cemented(h, min_class = "moderately", candidate_layers = NULL)
}
\arguments{
\item{h}{Horizons table.}

\item{min_class}{One of "weakly", "moderately", "strongly", "indurated".}

\item{candidate_layers}{Optional restriction.}
}
\description{
Used by petric variants (petrocalcic / petroduric / petrogypsic /
petroplinthic). The WRB 2022 ladder is: weakly < moderately <
strongly < indurated. Default threshold is "moderately".
}
\keyword{internal}
