% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-diagnostic-tests.R
\name{test_chernic_color}
\alias{test_chernic_color}
\title{Test for chroma <= 2 (moist) within the upper part of the profile}
\usage{
test_chernic_color(h, max_top_cm = 20, max_chroma = 2, candidate_layers = NULL)
}
\arguments{
\item{h}{Numeric threshold or option (see Details).}

\item{max_top_cm}{Numeric threshold or option (see Details).}

\item{max_chroma}{Numeric threshold or option (see Details).}

\item{candidate_layers}{Numeric threshold or option (see Details).}
}
\description{
Default upper boundary is 20 cm (Chernozem criterion: dark colour in
the upper 20 cm of the mollic horizon).
}
