% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-diagnostic-tests-v04.R
\name{test_coarse_fragments_above}
\alias{test_coarse_fragments_above}
\title{Test coarse-fragments percent above a threshold}
\usage{
test_coarse_fragments_above(h, min_pct = 90, max_top_cm = 25)
}
\description{
WRB 2022 alternative criterion for the leptic feature: coarse
fragments >= \code{min_pct} \% by volume in a layer within
\code{max_depth} of the surface. Used as an OR-alternative to the
R / Cr designation pattern.
}
\keyword{internal}
