% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-diagnostic-tests-v033.R
\name{test_starts_within}
\alias{test_starts_within}
\title{Test that a layer's top is at or below a target depth}
\usage{
test_starts_within(h, max_top_cm, candidate_layers = NULL)
}
\description{
Inverse of \code{\link{test_top_at_or_above}}: returns layers whose top
is shallower than or equal to \code{max_top_cm}, i.e. that start within
the upper part of the profile.
}
\keyword{internal}
