% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-diagnostic-tests-v033.R
\name{test_takyric_surface}
\alias{test_takyric_surface}
\title{Test for WRB 2022 Ch 3.2.15 takyric surface-crust signature}
\usage{
test_takyric_surface(h)
}
\description{
Surface mineral crust with: clay-loam-or-finer texture, platy/massive
structure, polygonal cracks >= 2 cm deep with spacing <= 20 cm,
rupture-resistance \\>= "hard" when dry, plasticity \\>= "moderately
plastic" when moist, EC < 4 dS/m OR >= 1 dS/m less than the layer
below. v0.3.3 enforces texture + structure + cracks + EC.
}
\keyword{internal}
