% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-diagnostic-tests.R
\name{test_texture_argic}
\alias{test_texture_argic}
\title{Test sandy-loam-or-finer texture (used by argic, ferralic)}
\usage{
test_texture_argic(h, candidate_layers = NULL)
}
\arguments{
\item{h}{Numeric threshold or option (see Details).}

\item{candidate_layers}{Numeric threshold or option (see Details).}
}
\description{
Test sandy-loam-or-finer texture (used by argic, ferralic)
}
