% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-andisols-usda.R
\name{thaptic_subgroup_usda}
\alias{thaptic_subgroup_usda}
\title{Thaptic Subgroup helper (Andisols)
Pass when, between 25 and 100 cm, a 10+ cm layer with OC > 3.0\%
and mollic colors exists, underlying lighter horizons.}
\usage{
thaptic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Thaptic Subgroup helper (Andisols)
Pass when, between 25 and 100 cm, a 10+ cm layer with OC > 3.0\%
and mollic colors exists, underlying lighter horizons.
}
