% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-spodosols-usda.R
\name{turbic_subgroup_usda}
\alias{turbic_subgroup_usda}
\title{Turbic Subgroup helper (Gelods)
Pass when gelic materials are present within 200 cm.
Implementation: cryoturbation + permafrost within 200 cm.}
\usage{
turbic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Turbic Subgroup helper (Gelods)
Pass when gelic materials are present within 200 cm.
Implementation: cryoturbation + permafrost within 200 cm.
}
