% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-spodosols-usda.R
\name{ultic_subgroup_usda}
\alias{ultic_subgroup_usda}
\title{Ultic Subgroup helper: argillic or kandic (any BS).}
\usage{
ultic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Ultic Subgroup helper: argillic or kandic (any BS).
}
