% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-ultisols-usda.R
\name{ultisol_qualifying_usda}
\alias{ultisol_qualifying_usda}
\title{Ultisol Order qualifier (USDA, KST 13ed, Ch 2)
Pass when argillic OR kandic horizon present + BS < 35\% in some
part of the upper 200 cm.}
\usage{
ultisol_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Ultisol Order qualifier (USDA, KST 13ed, Ch 2)
Pass when argillic OR kandic horizon present + BS < 35\% in some
part of the upper 200 cm.
}
