% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-spodosols-usda.R
\name{umbric_subgroup_usda}
\alias{umbric_subgroup_usda}
\title{Umbric Subgroup helper (in Spodosols)
Pass when umbric_epipedon_usda passes.}
\usage{
umbric_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Umbric Subgroup helper (in Spodosols)
Pass when umbric_epipedon_usda passes.
}
