% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-soilgrids.R
\name{utm_crs_for_point}
\alias{utm_crs_for_point}
\title{UTM zone EPSG code for a lon/lat point}
\usage{
utm_crs_for_point(lon, lat)
}
\description{
Picks the appropriate WGS84 UTM zone (32601..32660 northern,
32701..32760 southern) for a single coordinate. Used for metric
buffering.
}
\keyword{internal}
