% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-mollisols-usda.R
\name{vermic_subgroup_usda}
\alias{vermic_subgroup_usda}
\title{Vermic Subgroup helper (Vermudolls / Vermustolls)
Pass when worm_holes_pct >= 50\% in some horizon (KST 13ed worm
burrow criterion).}
\usage{
vermic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Vermic Subgroup helper (Vermudolls / Vermustolls)
Pass when worm_holes_pct >= 50\% in some horizon (KST 13ed worm
burrow criterion).
}
