% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-aridisols-usda.R
\name{vertic_aridisol_usda}
\alias{vertic_aridisol_usda}
\title{Vertic Aridisols helper -- delegates to vertic_subgroup_usda}
\usage{
vertic_aridisol_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Vertic Aridisols helper -- delegates to vertic_subgroup_usda
}
