% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-vertisols-usda.R
\name{vertisol_qualifying_usda}
\alias{vertisol_qualifying_usda}
\title{Vertisol Order qualifier (USDA, KST 13ed, Ch 2 / Ch 3 vertic horizon)
Pass when a vertic horizon (clay >= 30, cracks, slickensides, LE)
is present. Delegates to WRB \code{vertic_horizon}.}
\usage{
vertisol_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Vertisol Order qualifier (USDA, KST 13ed, Ch 2 / Ch 3 vertic horizon)
Pass when a vertic horizon (clay >= 30, cracks, slickensides, LE)
is present. Delegates to WRB \code{vertic_horizon}.
}
