% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-orders-usda.R
\name{vertisol_usda}
\alias{vertisol_usda}
\title{Vertisols (USDA Cap 16): slickensides + cracks.
Delegates to vertic_horizon.}
\usage{
vertisol_usda(pedon)
}
\description{
Vertisols (USDA Cap 16): slickensides + cracks.
Delegates to vertic_horizon.
}
