% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{vertissolo_ebanico}
\alias{vertissolo_ebanico}
\title{Vertissolos Ebanicos (Cap 17): caracter ebanico em B (cores escuras
dominantes).}
\usage{
vertissolo_ebanico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Vertissolos Ebanicos (Cap 17): caracter ebanico em B (cores escuras
dominantes).
}
