% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{vertissolo_haplico}
\alias{vertissolo_haplico}
\title{Vertissolos Haplicos (catch-all).}
\usage{
vertissolo_haplico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Vertissolos Haplicos (catch-all).
}
