% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{vertissolo_hidromorfico}
\alias{vertissolo_hidromorfico}
\title{Vertissolos Hidromorficos (Cap 17): horizonte glei OR caracter
redoxico.}
\usage{
vertissolo_hidromorfico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Vertissolos Hidromorficos (Cap 17): horizonte glei OR caracter
redoxico.
}
