% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-properties-wrb-v033.R
\name{vitric_properties}
\alias{vitric_properties}
\title{Vitric properties (WRB 2022 Ch 3.2.16)}
\usage{
vitric_properties(
  pedon,
  min_glass_pct = 5,
  min_alfe = 0.4,
  min_p_retention = 25
)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{min_glass_pct}{Numeric threshold or option (see Details).}

\item{min_alfe}{Numeric threshold or option (see Details).}

\item{min_p_retention}{Numeric threshold or option (see Details).}
}
\description{
Volcanic glass \\>= 5\% in 0.02-2 mm fraction, Al_ox + 1/2 Fe_ox \\>=
0.4\%, phosphate retention \\>= 25\%.
}
