% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-andisols-usda.R
\name{vitric_subgroup_usda}
\alias{vitric_subgroup_usda}
\title{Vitric Subgroup helper (Andisols)
Pass when volcanic_glass_pct >= 30 in a 25+ cm layer within 100 cm.}
\usage{
vitric_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Vitric Subgroup helper (Andisols)
Pass when volcanic_glass_pct >= 30 in a 25+ cm layer within 100 cm.
}
