% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-entisols-usda.R
\name{wassent_qualifying_usda}
\alias{wassent_qualifying_usda}
\title{Wassent Suborder qualifier (subaqueous Entisol).
Pass when site$water_table_cm_above_surface > 0 (water column
permanently above the surface).}
\usage{
wassent_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Wassent Suborder qualifier (subaqueous Entisol).
Pass when site$water_table_cm_above_surface > 0 (water column
permanently above the surface).
}
