% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-histosols-usda.R
\name{wassist_qualifying_usda}
\alias{wassist_qualifying_usda}
\title{Wassists Suborder qualifier (KST 13ed, Ch 10, p 203)}
\usage{
wassist_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\value{
A \code{\link{DiagnosticResult}}.
}
\description{
Histosols having a "field-observable water table 2 cm or more
above the soil surface for more than 21 hours of each day in
all years."  Diagnostic for the Wassists suborder.
}
\details{
Implementation: pass when \code{site$water_table_cm_above_surface}
is provided and >= 2 (positive = above surface).
}
