% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-oxisols-usda.R
\name{xanthic_subgroup_usda}
\alias{xanthic_subgroup_usda}
\title{Xanthic Subgroup helper (Oxisols)
Pass when 50\%+ colors have hue >= 7.5YR AND value >= 6 in B horizons.}
\usage{
xanthic_subgroup_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Xanthic Subgroup helper (Oxisols)
Pass when 50\%+ colors have hue >= 7.5YR AND value >= 6 in B horizons.
}
