% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{argissolo_vermelho_amarelo}
\alias{argissolo_vermelho_amarelo}
\title{Argissolos Vermelho-Amarelos (catch-all dos Argissolos)}
\usage{
argissolo_vermelho_amarelo(pedon)
}
\description{
Argissolos Vermelho-Amarelos (catch-all dos Argissolos)
}
