% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afsp.R
\name{benchmark_afsp}
\alias{benchmark_afsp}
\title{Benchmark soilKey WRB predictions against AfSP ground truth}
\usage{
benchmark_afsp(pedons, verbose = TRUE)
}
\arguments{
\item{pedons}{List of \code{\link{PedonRecord}} from
\code{\link{load_afsp_pedons}} or \code{\link{load_afsp_sample}}.}

\item{verbose}{Print progress.}
}
\value{
List with \code{accuracy}, \code{n_compared}, \code{confusion},
  \code{per_class_recall}.
}
\description{
Benchmark soilKey WRB predictions against AfSP ground truth
}
