% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-atributos-sibcs.R
\name{carater_salino}
\alias{carater_salino}
\title{Carater salino (SiBCS Cap 1, p 39): 4 <= CE < 7 dS/m.}
\usage{
carater_salino(pedon, min_ec = 4, max_ec = 7)
}
\description{
Carater salino (SiBCS Cap 1, p 39): 4 <= CE < 7 dS/m.
}
