% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-atributos-sibcs.R
\name{carater_sodico}
\alias{carater_sodico}
\title{Carater sodico (SiBCS Cap 1, p 39): saturacao por sodio (PST) >= 15%.}
\usage{
carater_sodico(pedon, min_pst = 15)
}
\description{
Carater sodico (SiBCS Cap 1, p 39): saturacao por sodio (PST) >= 15%.
}
