% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-atributos-sibcs.R
\name{contato_litico}
\alias{contato_litico}
\title{Contato litico (SiBCS Cap 1, p 40): rocha continua dura. Reuso de
\code{\link{continuous_rock}} via designacao R / Cr.}
\usage{
contato_litico(pedon)
}
\description{
Contato litico (SiBCS Cap 1, p 40): rocha continua dura. Reuso de
\code{\link{continuous_rock}} via designacao R / Cr.
}
