% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-atributos-sibcs.R
\name{contato_litico_fragmentario}
\alias{contato_litico_fragmentario}
\title{Contato litico fragmentario (SiBCS Cap 1, p 40): rocha fragmentada.}
\usage{
contato_litico_fragmentario(pedon)
}
\description{
Contato litico fragmentario (SiBCS Cap 1, p 40): rocha fragmentada.
}
