% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-orders-usda.R
\name{histosol_usda}
\alias{histosol_usda}
\title{Histosols (USDA Cap 10): organic materials.
Delegates to histic_horizon (WRB).}
\usage{
histosol_usda(pedon)
}
\description{
Histosols (USDA Cap 10): organic materials.
Delegates to histic_horizon (WRB).
}
