% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-horizons-sibcs.R
\name{horizonte_histico}
\alias{horizonte_histico}
\title{Horizonte hístico (SiBCS Cap 2, p 49-50)}
\usage{
horizonte_histico(pedon, min_oc_g_kg = 80)
}
\description{
Horizonte O ou H de coloracao preta/cinza muito escura/brunada,
\\>= 80 g/kg (8\\%) C organico, com:
\itemize{
  \item espessura \\>= 20 cm; OR
  \item espessura \\>= 40 cm se \\>= 75\\% volume tecido vegetal; OR
  \item espessura \\>= 10 cm sobre contato litico/fragmentario OR
        camada com \\>= 90\\% material > 2 mm.
}
}
