% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-inceptisols-usda.R
\name{humic_inceptisol_usda}
\alias{humic_inceptisol_usda}
\title{Humic Inceptisol Suborder helper (Hum*)
Pass when umbric or mollic epipedon present + thick (>= 25 cm).}
\usage{
humic_inceptisol_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Humic Inceptisol Suborder helper (Hum*)
Pass when umbric or mollic epipedon present + thick (>= 25 cm).
}
