% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-ultisols-usda.R
\name{humult_qualifying_usda}
\alias{humult_qualifying_usda}
\title{Humult Suborder qualifier (Ultisols with thick humus accumulation)
Pass when 0.9\% OC weighted average in 0-15 cm AND/OR
organic carbon mass >= 12 kg/m2 in 0-100 cm (proxy via humic_oxisol_usda
with lower threshold).}
\usage{
humult_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Humult Suborder qualifier (Ultisols with thick humus accumulation)
Pass when 0.9\% OC weighted average in 0-15 cm AND/OR
organic carbon mass >= 12 kg/m2 in 0-100 cm (proxy via humic_oxisol_usda
with lower threshold).
}
