% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-materials-wrb-v033.R
\name{hyposulfidic_material}
\alias{hyposulfidic_material}
\title{Hyposulfidic material (WRB 2022 Ch 3.3.9): same S and pH as
hypersulfidic but does NOT consist of hypersulfidic (i.e. not capable
of severe acidification). v0.3.3: returns sulfidic layers that don't
meet hypersulfidic.}
\usage{
hyposulfidic_material(pedon, min_s_pct = 0.01, min_pH = 4)
}
\description{
Hyposulfidic material (WRB 2022 Ch 3.3.9): same S and pH as
hypersulfidic but does NOT consist of hypersulfidic (i.e. not capable
of severe acidification). v0.3.3: returns sulfidic layers that don't
meet hypersulfidic.
}
