% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fixtures-sibcs.R
\name{make_cambissolo_canonical}
\alias{make_cambissolo_canonical}
\title{Perfil canonico de Cambissolo (SiBCS 5a ed., Cap 6)}
\usage{
make_cambissolo_canonical()
}
\description{
Reusa fixture WRB Cambisol -- B incipiente sem ser plintico,
vertico, planico, etc.
}
