% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fixtures-sibcs.R
\name{make_gleissolo_canonical}
\alias{make_gleissolo_canonical}
\title{Perfil canonico de Gleissolo (SiBCS 5a ed., Cap 9)}
\usage{
make_gleissolo_canonical()
}
\description{
Reusa fixture WRB Gleysol -- horizonte glei dentro de 50 cm.
}
