% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fixtures-sibcs.R
\name{make_planossolo_canonical}
\alias{make_planossolo_canonical}
\title{Perfil canonico de Planossolo (SiBCS 5a ed., Cap 15)}
\usage{
make_planossolo_canonical()
}
\description{
Solo com horizonte E sobrejacente a B planico (mudanca textural
abrupta + cores neutras + cromas baixos).
}
