% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{neossolo_litolico}
\alias{neossolo_litolico}
\title{Neossolos Litolicos (Cap 12): contato litico ou litico fragmentario
\\<= 50 cm.}
\usage{
neossolo_litolico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Neossolos Litolicos (Cap 12): contato litico ou litico fragmentario
\\<= 50 cm.
}
