% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{neossolo_quartzarenico}
\alias{neossolo_quartzarenico}
\title{Neossolos Quartzarenicos (Cap 12): textura areia/areia franca em
todos os horizontes ate 150 cm + 95\% quartzo.}
\usage{
neossolo_quartzarenico(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Neossolos Quartzarenicos (Cap 12): textura areia/areia franca em
todos os horizontes ate 150 cm + 95\% quartzo.
}
