% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-subordens-sibcs.R
\name{organossolo_folico}
\alias{organossolo_folico}
\title{Organossolos Folicos (Cap 14): horizonte O histico (drenado).
Detectado via designation pattern \"^O\".}
\usage{
organossolo_folico(pedon)
}
\description{
Organossolos Folicos (Cap 14): horizonte O histico (drenado).
Detectado via designation pattern \"^O\".
}
