% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-horizons-wrb-v033.R
\name{petroduric}
\alias{petroduric}
\title{Petroduric horizon (WRB 2022): cemented duric.}
\usage{
petroduric(pedon, min_thickness = 10, min_duripan_pct = 10)
}
\description{
Petroduric horizon (WRB 2022): cemented duric.
}
