% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-horizons-wrb-v033.R
\name{petrogypsic}
\alias{petrogypsic}
\title{Petrogypsic horizon (WRB 2022): cemented gypsic.}
\usage{
petrogypsic(pedon, min_thickness = 10, min_gypsum_pct = 5)
}
\description{
Petrogypsic horizon (WRB 2022): cemented gypsic.
}
