% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-horizons-wrb-v033.R
\name{petroplinthic}
\alias{petroplinthic}
\title{Petroplinthic horizon (WRB 2022): cemented plinthic.}
\usage{
petroplinthic(pedon, min_thickness = 10, min_plinthite_pct = 15)
}
\description{
Petroplinthic horizon (WRB 2022): cemented plinthic.
}
