% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_arenicolic}
\alias{qual_arenicolic}
\title{Arenicolic supplementary qualifier (an): faunal sand burrows}
\usage{
qual_arenicolic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
WRB 2022 Ch 5: "Containing layers with extensive sand-grade
bioturbation (faunal burrows from earthworms / ants / termites)."
Implementation: \code{bioturbation_density} \\>= "common".
}
