% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0964.R
\name{qual_biocrustic}
\alias{qual_biocrustic}
\title{Biocrustic supplementary qualifier (bk): biological soil crust}
\usage{
qual_biocrustic(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
WRB 2022 Ch 5: "Surface biological crust (cyanobacteria, algae,
lichens, mosses)." Implementation: \code{surface_crust_type} matching
biological pattern in upper 5 cm.
}
