% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022.R
\name{qual_dystric}
\alias{qual_dystric}
\title{Dystric qualifier (dy): low base saturation throughout. v0.9: BS <
50\% from 20 to 100 cm in mineral material.}
\usage{
qual_dystric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Dystric qualifier (dy): low base saturation throughout. v0.9: BS <
50\% from 20 to 100 cm in mineral material.
}
