% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v0963.R
\name{qual_epidystric}
\alias{qual_epidystric}
\title{Epidystric supplementary qualifier (epd): dystric only in upper 50 cm}
\usage{
qual_epidystric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
WRB 2022 Ch 5: "Dystric (BS < 50\%) in upper 50 cm and eutric below."
}
