% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091c.R
\name{qual_hyperdystric}
\alias{qual_hyperdystric}
\title{Hyperdystric qualifier (yd): base saturation < 5\% throughout the
upper 100 cm (mineral soil layers only). Stronger than Dystric (BS
< 50\%).}
\usage{
qual_hyperdystric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Hyperdystric qualifier (yd): base saturation < 5\% throughout the
upper 100 cm (mineral soil layers only). Stronger than Dystric (BS
< 50\%).
}
