% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifiers-wrb2022-v091c.R
\name{qual_hypereutric}
\alias{qual_hypereutric}
\title{Hypereutric qualifier (ye): base saturation >= 80\% throughout the
upper 100 cm. Stronger than Eutric (BS >= 50\%).}
\usage{
qual_hypereutric(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Hypereutric qualifier (ye): base saturation >= 80\% throughout the
upper 100 cm. Stronger than Eutric (BS >= 50\%).
}
