% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-entisols-usda.R
\name{quartzipsamment_qualifying_usda}
\alias{quartzipsamment_qualifying_usda}
\title{Quartzipsamment helper (Quartzipsamments: >= 95\% silica)
v0.8 proxy: clay <= 5\% AND coarse_fragments_pct <= 5\%.}
\usage{
quartzipsamment_qualifying_usda(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Quartzipsamment helper (Quartzipsamments: >= 95\% silica)
v0.8 proxy: clay <= 5\% AND coarse_fragments_pct <= 5\%.
}
