% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-materials-wrb-v033.R
\name{technic_hard_material}
\alias{technic_hard_material}
\title{Technic hard material (WRB 2022 Ch 3.3.18): consolidated human-made
material (asphalt, concrete, worked stones).}
\usage{
technic_hard_material(pedon)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}
}
\description{
Technic hard material (WRB 2022 Ch 3.3.18): consolidated human-made
material (asphalt, concrete, worked stones).
}
