% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-diagnostic-tests.R
\name{test_ferralic_texture}
\alias{test_ferralic_texture}
\title{Ferralic texture: sandy loam or finer (same predicate as argic)}
\usage{
test_ferralic_texture(h, candidate_layers = NULL)
}
\arguments{
\item{h}{Numeric threshold or option (see Details).}

\item{candidate_layers}{Numeric threshold or option (see Details).}
}
\description{
Ferralic texture: sandy loam or finer (same predicate as argic)
}
