% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-horizons-wrb-v033.R
\name{thionic}
\alias{thionic}
\title{Thionic horizon (WRB 2022): post-oxidation acid sulfate horizon.
Requires sulfidic_s_pct >= 0.01 AND pH(H2O) <= 4.}
\usage{
thionic(pedon, min_thickness = 15, max_pH = 4, min_sulfidic_s = 0.01)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{min_thickness}{Numeric threshold or option (see Details).}

\item{max_pH}{Numeric threshold or option (see Details).}

\item{min_sulfidic_s}{Numeric threshold or option (see Details).}
}
\description{
Thionic horizon (WRB 2022): post-oxidation acid sulfate horizon.
Requires sulfidic_s_pct >= 0.01 AND pH(H2O) <= 4.
}
