% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-orders-usda.R
\name{ultisol_usda}
\alias{ultisol_usda}
\title{Ultisols (USDA Cap 15): argillic/kandic horizon + base saturation < 35%.}
\usage{
ultisol_usda(pedon)
}
\description{
Ultisols (USDA Cap 15): argillic/kandic horizon + base saturation < 35%.
}
