% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-andisols-usda.R
\name{vitrand_qualifying_usda}
\alias{vitrand_qualifying_usda}
\title{Vitrands qualifier (Cap 6, pp 117-118)
Pass when 1500 kPa water retention < 15\% (air-dried) and
< 30\% (undried) throughout 60\%+ of the thickness. v0.8 proxy:
uses water_content_1500kpa < 15\%.}
\usage{
vitrand_qualifying_usda(pedon, max_top_cm = 60)
}
\arguments{
\item{pedon}{A \code{\link{PedonRecord}}.}

\item{max_top_cm}{Numeric threshold or option (see Details).}
}
\description{
Vitrands qualifier (Cap 6, pp 117-118)
Pass when 1500 kPa water retention < 15\% (air-dried) and
< 30\% (undried) throughout 60\%+ of the thickness. v0.8 proxy:
uses water_content_1500kpa < 15\%.
}
