% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-properties-wrb-v033.R
\name{yermic_properties}
\alias{yermic_properties}
\title{Yermic properties (WRB 2022 Ch 3.2.17) -- per-pedon test wrapping
\code{\link{test_yermic_surface}}.}
\usage{
yermic_properties(pedon)
}
\description{
Yermic properties (WRB 2022 Ch 3.2.17) -- per-pedon test wrapping
\code{\link{test_yermic_surface}}.
}
